### 2024-11-12 #7459 @editorialbot
[PRE REVIEW]: svZeroDSolver: A modular package for lumped-parameter cardiovascular simulations
**Submitting author:** @menon-karthik (<a href="http://orcid.org/0000-0002-7494-7585">Karthik Menon</a>)
**Repository:** https://github.com/SimVascular/svZeroDSolver.git
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.0
