### 2024-12-09 #7459 @philipcardiff
Hello @TS-CUBED, Would you be interested in reviewing the submission “svZeroDSolver: A modular package for lumped-parameter cardiovascular simulations” which we received at JOSS? The submission link is https://github.com/openjournals/joss-reviews/issues/7459, and the repository is at https://github.com/SimVascular/svZeroDSolver.git.

Given your work on https://github.com/TS-CUBED/CirculatorySystemModels.jl, this submission may be of interest to you.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. To get an idea of what is expected, you can look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html).
