### 2024-12-11 #7459 @TS-CUBED
Hello Phil,

Yes, happy  to do that. What is the expected timeframe?

Best wishes,

Torsten

–

Dr Torsten Schenkel
Dr.-Ing. habil. CEng FIMechE

Associate Professor of Continuum Mechanics

Sheffield Hallam University
School of Engineering and Built Environment
College of Business, Technology and Engineering
Room 4206, Sheaf Building

Sheffield, S1 1WB
UK

Tel +44 (0)114 225 6294
Profile: https://www.shu.ac.uk/about-us/our-people/staff-profiles/torsten-schenkel
EC2M3 Lab: https://blogs.shu.ac.uk/ecm<https://blogs.shu.ac.uk/ecm/>
Github: https://github.com/TS-CUBED
Github Pages: https://ts-cubed.github.io<https://ts-cubed.github.io/>
