### 2024-12-11 #7459 @philipcardiff
Thanks, Torsten. Once the review starts, you will be presented with a reviewer checklist, which you would ideally start on within the first 3-4 weeks. Unlike a traditional academic review, for JOSS, you can incrementally make progress on the review, providing feedback to the authors if appropriate. If anything is unclear, feel free to ask.
