### 2024-12-17 #7595 @editorialbot
[REVIEW]: svZeroDSolver: A modular package for lumped-parameter cardiovascular simulations
**Submitting author:** @menon-karthik (<a href="http://orcid.org/0000-0002-7494-7585">Karthik Menon</a>)
**Repository:** https://github.com/SimVascular/svZeroDSolver.git
**Branch with paper.md** (empty if default branch): paper
**Version:** v3.0
