### 2024-12-17 #7459 @philipcardiff
We can now move to https://github.com/openjournals/joss-reviews/issues/7595 for the review
