### 2024-12-17 #7595 @TS-CUBED
The license follows the MIT license, but does not specify it. It may be useful to include the name "MIT License" in the License file to make it clear to the user that it is using this common license.
