### 2025-01-23 #7595 @marchirschvogel
I agree with @TS-CUBED that the license is of MIT-type, but should probably state this explicitly in the license file.
