### 2025-01-23 #7595 @marchirschvogel
I do not have any conflict of interest with this work. Nonetheless, for the sake of transparency, I mention that there is a recent publication where I and some of the authors have contributed to: https://doi.org/10.1016/j.cma.2024.117485. However, this was independent work at different institutions without any type of direct collaboration.
