### 2025-01-26 SimVascular/svZeroDSolver#143 @marchirschvogel
[JOSS review] Error on installation with Python 3.10
### Problem

This is related to https://github.com/openjournals/joss-reviews/issues/7595

Running `python3.10 -m pip install git+https://github.com/simvascular/svZeroDSolver.git` under my system (Ubuntu 22.04.5 LTS, Python 3.10.12, cmake version 3.22.1) gives the following error:

Defaulting to user installation because normal site-packages is not writeable
Collecting git+https://github.com/simvascular/svZeroDSolver.git
  Cloning https://github.com/simvascular/svZeroDSolver.git to /tmp/pip-req-build-pgdb_2c_
  Running command git clone --filter=blob:none --quiet https://github.com/simvascular/svZeroDSolver.git /tmp/pip-req-build-pgdb_2c_
  Resolved https://github.com/simvascular/svZeroDSolver.git to commit 8a58d680f3431c0e1cddda52cdec0e57d3ced254
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting pandas (from pysvzerod==2.0)
  Downloading pandas-2.2.3-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (89 kB)
Requirement already satisfied: numpy in ./.local/lib/python3.10/site-packages (from pysvzerod==2.0) (2.1.3)
Requirement already satisfied: python-dateutil>=2.8.2 in ./.local/lib/python3.10/site-packages (from pandas->pysvzerod==2.0) (2.9.0.post0)
Requirement already satisfied: pytz>=2020.1 in /usr/lib/python3/dist-packages (from pandas->pysvzerod==2.0) (2022.1)
Collecting tzdata>=2022.7 (from pandas->pysvzerod==2.0)
  Downloading tzdata-2025.1-py2.py3-none-any.whl.metadata (1.4 kB)
Requirement already satisfied: six>=1.5 in /usr/lib/python3/dist-packages (from python-dateutil>=2.8.2->pandas->pysvzerod==2.0) (1.16.0)
Downloading pandas-2.2.3-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (13.1 MB)
Downloading tzdata-2025.1-py2.py3-none-any.whl (346 kB)
Building wheels for collected packages: pysvzerod
  Building wheel for pysvzerod (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for pysvzerod (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [110 lines of output]
      running bdist_wheel
      running build
      running build_py
      creating build/lib.linux-x86_64-cpython-310/tests
      copying tests/test_calibrator.py -> build/lib.linux-x86_64-cpython-310/tests
      copying tests/test_dirgraph.py -> build/lib.linux-x86_64-cpython-310/tests
      copying tests/conftest.py -> build/lib.linux-x86_64-cpython-310/tests
      copying tests/utils.py -> build/lib.linux-x86_64-cpython-310/tests
      copying tests/__init__.py -> build/lib.linux-x86_64-cpython-310/tests
      copying tests/test_io.py -> build/lib.linux-x86_64-cpython-310/tests
      copying tests/compute_ref_sol.py -> build/lib.linux-x86_64-cpython-310/tests
      copying tests/test_solver.py -> build/lib.linux-x86_64-cpython-310/tests
      running build_ext
      -- The C compiler identification is GNU 11.4.0
      -- The CXX compiler identification is GNU 11.4.0
      -- Detecting C compiler ABI info
      -- Detecting C compiler ABI info - done
      -- Check for working C compiler: /usr/bin/cc - skipped
      -- Detecting C compile features
      -- Detecting C compile features - done
      -- Detecting CXX compiler ABI info
      -- Check for working CXX compiler: /usr/bin/c++ - skipped
      -- Detecting CXX compile features
      --
      -- Configured Eigen 3.4.90
      --
      -- pybind11 v2.14.0 dev1
      -- Found PythonInterp: /usr/bin/python3.8 (found suitable version "3.8.10", minimum required is "3.8")
      -- Found PythonLibs: python3.8
      -- Performing Test HAS_FLTO
      -- Performing Test HAS_FLTO - Success
      -- Using the multi-header code from /tmp/pip-req-build-pgdb_2c_/build/temp.linux-x86_64-cpython-310/_deps/json-src/include/
      -- Configuring done
      CMake Error in CMakeLists.txt:
        Imported target "pybind11::module" includes non-existent path

          "/usr/include/python3.8"

        in its INTERFACE_INCLUDE_DIRECTORIES.  Possible reasons include:

        * The path was deleted, renamed, or moved to another location.

        * An install or uninstall procedure did not complete successfully.

        * The installation package was faulty and references files it does not
        provide.

      -- Generating done
      CMake Generate step failed.  Build files cannot be regenerated correctly.
      Traceback (most recent call last):
        File "/home/mh/.local/lib/python3.10/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
        File "/home/mh/.local/lib/python3.10/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
        File "/home/mh/.local/lib/python3.10/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 251, in build_wheel
          return _build_backend().build_wheel(wheel_directory, config_settings,
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/build_meta.py", line 438, in build_wheel
          return _build(['bdist_wheel', '--dist-info-dir', str(metadata_directory)])
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/build_meta.py", line 426, in _build
          return self._build_with_temp_dir(
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/build_meta.py", line 407, in _build_with_temp_dir
          self.run_setup()
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/build_meta.py", line 320, in run_setup
          exec(code, locals())
        File "<string>", line 4, in <module>
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/__init__.py", line 117, in setup
          return distutils.core.setup(**attrs)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/core.py", line 186, in setup
          return run_commands(dist)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/core.py", line 202, in run_commands
          dist.run_commands()
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/dist.py", line 983, in run_commands
          self.run_command(cmd)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/dist.py", line 999, in run_command
          super().run_command(command)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/dist.py", line 1002, in run_command
          cmd_obj.run()
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/command/bdist_wheel.py", line 379, in run
          self.run_command("build")
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/cmd.py", line 339, in run_command
          self.distribution.run_command(command)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/dist.py", line 999, in run_command
          super().run_command(command)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/dist.py", line 1002, in run_command
          cmd_obj.run()
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/command/build.py", line 136, in run
          self.run_command(cmd_name)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/cmd.py", line 339, in run_command
          self.distribution.run_command(command)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/dist.py", line 999, in run_command
          super().run_command(command)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/dist.py", line 1002, in run_command
          cmd_obj.run()
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/command/build_ext.py", line 99, in run
          _build_ext.run(self)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/command/build_ext.py", line 365, in run
          self.build_extensions()
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/command/build_ext.py", line 481, in build_extensions
          self._build_extensions_serial()
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/setuptools/_distutils/command/build_ext.py", line 507, in _build_extensions_serial
          self.build_extension(ext)
        File "/tmp/pip-build-env-bsuyiy1y/overlay/local/lib/python3.10/dist-packages/cmake_setuptools/cmake/__init__.py", line 54, in build_extension
          subprocess.check_call(cmake_args,
        File "/usr/lib/python3.10/subprocess.py", line 369, in check_call
          raise CalledProcessError(retcode, cmd)
      subprocess.CalledProcessError: Command '['/usr/bin/cmake', '/tmp/pip-req-build-pgdb_2c_', '-DCMAKE_LIBRARY_OUTPUT_DIRECTORY=/tmp/pip-req-build-pgdb_2c_/build/lib.linux-x86_64-cpython-310', '-DCMAKE_BUILD_TYPE=Release']' returned non-zero exit status 1.
      [end of output]

  ERROR: Failed building wheel for pysvzerod
Failed to build pysvzerod
ERROR: ERROR: Failed to build installable wheels for some pyproject.toml based projects (pysvzerod)

### Solution

I would expect that installation runs error-free, particularly since I haven't found special system requirements.

### Additional context

_No response_

### Code of Conduct

- [x] I agree to follow this project's Code of Conduct and Contributing Guidelines
