### 2025-01-26 #7595 @marchirschvogel
I'm having issues with the Python installation and opened a respective issue: https://github.com/SimVascular/svZeroDSolver/issues/143.
