### 2025-01-26 #7595 @marchirschvogel
I'm missing a "State of the field" section, or at least references to other open source projects that support 0D vascular modeling. Maybe at least some of the contributions that were mentioned in https://github.com/openjournals/joss-reviews/issues/7459#issuecomment-2470139393 could discussed in terms of commonalities/differences to svZeroDSolver.
