### 2025-02-06 SimVascular/svZeroDSolver#143 @mrp089
I think there might be an issue with your system. You're requesting `python3.10`, but it seems pybind11 finds `python3.8` on your system. What happens if you don't specify a Python version during installation?
