### 2025-02-07 SimVascular/svZeroDSolver#143 @marchirschvogel
This does not help, using `pip install git+https://github.com/simvascular/svZeroDSolver.git` produces the same error. In CMakeLists.txt l. 97, it seems to hard-set a pybind11 version related to python3.8 if I understand this correctly:
`set(PYBIND11_PYTHON_VERSION 3.8 CACHE STRING "")`.
