+++ src/model/BloodVesselNew.cpp
@@ -1,0 +1,131 @@
+// Copyright (c) Stanford University, The Regents of the University of
+//               California, and others.
+//
+// All Rights Reserved.
+//
+// See Copyright-SimVascular.txt for additional details.
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject
+// to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included
+// in all copies or substantial portions of the Software.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+// TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+// OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+#include "BloodVesselNew.h"
+
+void BloodVesselNew::setup_dofs(DOFHandler &dofhandler) {
+  Block::setup_dofs_(dofhandler, 2, {});
+}
+
+void BloodVesselNew::update_constant(SparseSystem &system,
+                                  std::vector<double> &parameters) {
+  // Get parameters
+  double capacitance = parameters[global_param_ids[ParamId::CAPACITANCE]];
+  double inductance = parameters[global_param_ids[ParamId::INDUCTANCE]];
+  double resistance = parameters[global_param_ids[ParamId::RESISTANCE]];
+
+  // Set element contributions
+  // coeffRef args are the indices (i,j) of the matrix
+  // global_eqn_ids: number of rows in the matrix, set in setup_dofs
+  // global_var_ids: number of columns, organized as pressure and flow of all
+  // inlets and then all outlets of the block
+  system.E.coeffRef(global_eqn_ids[0], global_var_ids[3]) = -inductance;
+  system.E.coeffRef(global_eqn_ids[1], global_var_ids[0]) = -capacitance;
+  system.E.coeffRef(global_eqn_ids[1], global_var_ids[1]) =
+      capacitance * resistance;
+  system.F.coeffRef(global_eqn_ids[0], global_var_ids[0]) = 1.0;
[…]
