### 2025-02-24 3a43dae
added test case for BloodVesselNew block
