### 2025-02-24 dce4e3a
updated test_solver.py to include new BVN testcase
