### 2025-02-26 #7595 @menon-karthik
@philipcardiff Not currently, thanks for checking. Sorry I have been slow to get back to this because I just moved to a new position. Does JOSS have specific timeframes to respond to reviewers?
