### 2025-03-04 #7595 @philipcardiff
> [@philipcardiff](https://github.com/philipcardiff) Not currently, thanks for checking. Sorry I have been slow to get back to this because I just moved to a new position. Does JOSS have specific timeframes to respond to reviewers?

Hi @menon-karthik,

No problem at all, and congratulations on your new position! 🎉

JOSS is quite flexible with timelines, but we generally encourage authors to respond to reviewers within a few weeks to keep the process moving. That said, if you need more time, that can be accommodated, as there is no strict deadline.
