### 2025-03-11 f2b4e8a
Added test cases and documentation
