+++ tests/test_dirgraph.py
@@ -19,7 +19,10 @@ excluded_files = [
     'closedLoopHeart_singleVessel.json',
     'closedLoopHeart_withCoronaries.json',
     'coupledBlock_closedLoopHeart_singleVessel.json',
-    'coupledBlock_closedLoopHeart_withCoronaries.json'
+    'coupledBlock_closedLoopHeart_withCoronaries.json',
+    'closedLoopHeart_singleVessel_mistmatchPeriod.json',
+    'pulsatileFlow_R_RCR_mismatchPeriod.json',
+    'pulsatileFlow_CStenosis_steadyPressure_definedPeriod.json'
 ]
 
 # Generate the list of JSON files to test
