+++ .github/workflows/documentation.yml
@@ -18,7 +18,7 @@ jobs:
           doxyfile-path: 'docs/Doxyfile'
           enable-latex: true
       - name: Save documentation
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: documentation
           path: ./docs/build/html
