### 2025-03-11 c3a160c
Clang fix
