+++ paper/paper.bib
@@ -79,7 +79,8 @@ url = {https://doi.org/10.1007/s10439-010-0083-6}
   volume={18},
   pages={503--529},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10237-018-1098-4}
 }
 
 @article{menon2023predictors,
@@ -99,7 +100,8 @@ url = {https://doi.org/10.1007/s10439-010-0083-6}
   volume={40},
   pages={2228--2242},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10439-012-0579-3}
 }
 
 @article{kerckhoffs2007coupling,
@@ -109,7 +111,8 @@ url = {https://doi.org/10.1007/s10439-010-0083-6}
   volume={35},
   pages={1--18},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10439-006-9212-7}
 }
 
 @article{menon2024cardiovascular, 
@@ -178,7 +181,7 @@ pages = {116414},
 year = {2023},
 note = {A Special Issue in Honor of the Lifetime Achievements of T. J. R. Hughes},
 issn = {0045-7825},
-doi = {https://doi.org/10.1016/j.cma.2023.116414},
+doi = {10.1016/j.cma.2023.116414},
 url = {https://www.sciencedirect.com/science/article/pii/S0045782523005388},
 author = {Charles A. Taylor and Kersten Petersen and Nan Xiao and Matthew Sinclair and Ying Bai and Sabrina R. Lynch and Adam UpdePac and Michiel Schaap},
 keywords = {Patient-specific modeling, Blood flow, Deep learning, Coronary artery},
@@ -200,7 +203,7 @@ year = {2017}
 
 @article{Lee2024,
 author = {Lee, John D and Richter, Jakob and Pfaller, Martin R and Szafron, Jason M and Menon, Karthik and Zanoni, Andrea and Ma, Michael R and Feinstein, Jeffrey A and Kreutzer, Jacqueline and Marsden, Alison L and Schiavazzi, Daniele E},
-doi = {https://doi.org/10.1002/cnm.3820},
+doi = {10.1002/cnm.3820},
 journal = {International Journal for Numerical Methods in Biomedical Engineering},
 keywords = { neural network emulator, parameterization of complex interventional repairs,digital twin for interventional planning},
 number = {5},
@@ -213,7 +216,7 @@ year = {2024}
 
 @article{zanoni2024,
 author = {Zanoni, Andrea and Geraci, Gianluca and Salvador, Matteo and Menon, Karthik and Marsden, Alison L and Schiavazzi, Daniele E},
-doi = {https://doi.org/10.1016/j.cma.2024.117119},
+doi = {10.1016/j.cma.2024.117119},
 issn = {0045-7825},
 journal = {Computer Methods in Applied Mechanics and Engineering},
 keywords = { Active subspaces, Autoencoders, Monte Carlo estimators, Normalizing flows, Uncertainty quantification,Multifidelity},
@@ -245,18 +248,21 @@ pages = {449-476},
 year = {2024},
 series = {Biomechanics of Living Organs},
 isbn = {978-0-323-95484-6},
-doi = {https://doi.org/10.1016/B978-0-323-95484-6.00016-6},
+doi = {10.1016/B978-0-323-95484-6.00016-6},
 url = {https://www.sciencedirect.com/science/article/pii/B9780323954846000166},
 author = {Martin R. Pfaller and Luca Pegolotti and Jonathan Pham and Natalia L. Rubio and Alison L. Marsden},
 }
 
 @article{rubio2024,
[…]
