+++ paper/paper.bib
@@ -250,7 +250,7 @@ series = {Biomechanics of Living Organs},
 isbn = {978-0-323-95484-6},
 doi = {10.1016/B978-0-323-95484-6.00016-6},
 url = {https://www.sciencedirect.com/science/article/pii/B9780323954846000166},
-author = {Martin R. Pfaller and Luca Pegolotti and Jonathan Pham and Natalia L. Rubio and Alison L. Marsden},
+author = {Martin R. Pfaller and Luca Pegolotti and Jonathan Pham and Natalia L. Rubio and Alison L. Marsden}
 }
 
 @article{rubio2024,
@@ -262,7 +262,7 @@ year = {2025},
 issn = {0010-4825},
 doi = {10.1016/j.compbiomed.2024.109420},
 url = {https://www.sciencedirect.com/science/article/pii/S0010482524015051},
-author = {Natalia L. Rubio and Luca Pegolotti and Martin R. Pfaller and Eric F. Darve and Alison L. Marsden},
+author = {Natalia L. Rubio and Luca Pegolotti and Martin R. Pfaller and Eric F. Darve and Alison L. Marsden}
 }
 
 @article{Vignon-Clementel2006,
@@ -299,3 +299,33 @@ year = {2019}
       journal={arXiv},
       url={https://arxiv.org/abs/2409.02247},
 }
+
+@article{AFRICA2024109039,
+title = {lifex-cfd: An open-source computational fluid dynamics solver for cardiovascular applications},
+journal = {Computer Physics Communications},
+volume = {296},
+pages = {109039},
+year = {2024},
+issn = {0010-4655},
+doi = {10.1016/j.cpc.2023.109039},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465523003843},
+author = {Pasquale Claudio Africa and Ivan Fumagalli and Michele Bucelli and Alberto Zingaro and Marco Fedele and Luca Dede' and Alfio Quarteroni},
+}
+
+@article{crimson2021,
+    doi = {10.1371/journal.pcbi.1008881},
+    author = {Arthurs, Christopher J. AND Khlebnikov, Rostislav AND Melville, Alex AND Marčan, Marija AND Gomez, Alberto AND Dillon-Murphy, Desmond AND Cuomo, Federica AND Silva Vieira, Miguel AND Schollenberger, Jonas AND Lynch, Sabrina R. AND Tossas-Betancourt, Christopher AND Iyer, Kritika AND Hopper, Sara AND Livingston, Elizabeth AND Youssefi, Pouya AND Noorani, Alia AND Ben Ahmed, Sabrina AND Nauta, Foeke J. H. AND van Bakel, Theodorus M. J. AND Ahmed, Yunus AND van Bakel, Petrus A. J. AND Mynard, Jonathan AND Di Achille, Paolo AND Gharahi, Hamid AND Lau, Kevin D. AND Filonova, Vasilina AND Aguirre, Miquel AND Nama, Nitesh AND Xiao, Nan AND Baek, Seungik AND Garikipati, Krishna AND Sahni, Onkar AND Nordsletten, David AND Figueroa, C. Alberto},
+    journal = {PLOS Computational Biology},
+    publisher = {Public Library of Science},
+    title = {CRIMSON: An open-source software framework for cardiovascular integrated modelling and simulation},
+    year = {2021},
+    month = {05},
+    volume = {17},
+    url = {https://doi.org/10.1371/journal.pcbi.1008881},
+    pages = {1-21},
+    number = {5},
+}
+
[…]
