### 2025-03-11 SimVascular/svZeroDSolver#143 @menon-karthik
@marchirschvogel We cannot reproduce the error. Can you please try using a virtual environment to install and report back if you have the same error? We generally install using a virtual environment, and if that resolves the error, we can update the docs to include using a virtual environment.
