+++ .github/workflows/test.yml
@@ -16,8 +16,7 @@ jobs:
           auto-update-conda: true
       - name: Install ubuntu dependencies
         if: startsWith(matrix.os, 'ubuntu')
-        run: sudo apt update && sudo apt install build-essential cmake lcov
-        run: sudo apt install --only-upgrade libstdc++6
+        run: sudo apt update && sudo apt install build-essential cmake lcov && sudo apt install --only-upgrade libstdc++6
       - name: Install svZeroDSolver
         run: |
           #export PATH="/usr/share/miniconda/bin:$PATH"
