### 2025-03-11 e0b661f
try conda forge libstdc++ 2
