+++ .github/workflows/test.yml
@@ -16,15 +16,16 @@ jobs:
           auto-update-conda: true
       - name: Install ubuntu dependencies
         if: startsWith(matrix.os, 'ubuntu')
-        run: sudo apt update && sudo apt install build-essential cmake lcov && strings /usr/lib/x86_64-linux-gnu/libstdc++.so.6 
+        run: sudo apt update && sudo apt install build-essential cmake lcov && strings /usr/lib/x86_64-linux-gnu/libstdc++.so.6 | grep GLIBCXX
       - name: Install svZeroDSolver
         run: |
           #export PATH="/usr/share/miniconda/bin:$PATH"
           #alias conda="$CONDA/bin/conda"
           conda create -n zerod python=3.11.4
-          grep GLIBCXX && strings /usr/share/miniconda/envs/zerod/bin/../lib/libstdc++.so.6 | grep GLIBCXX
-          conda run -n zerod install -c conda-forge libstdcxx-ng --update-deps
-          grep GLIBCXX && strings /usr/share/miniconda/envs/zerod/bin/../lib/libstdc++.so.6 | grep GLIBCXX
+          strings /usr/share/miniconda/envs/zerod/bin/../lib/libstdc++.so.6 | grep GLIBCXX
+          #conda run -n zerod install -c conda-forge libstdcxx-ng --update-deps
+          conda run -n zerod install -c conda-forge libstdcxx-ng
+          strings /usr/share/miniconda/envs/zerod/bin/../lib/libstdc++.so.6 | grep GLIBCXX
           conda run -n zerod pip install -e ".[dev]"
       - name: Install Networkx
         run: |
