### 2025-03-11 8a8ce01
cleaned up test.yml
