+++ .github/workflows/test.yml
@@ -16,22 +16,17 @@ jobs:
           auto-update-conda: true
       - name: Install ubuntu dependencies
         if: startsWith(matrix.os, 'ubuntu')
-        run: sudo apt update && sudo apt install build-essential cmake lcov && strings /usr/lib/x86_64-linux-gnu/libstdc++.so.6 | grep GLIBCXX
-      - name: Install svZeroDSolver
+        run: sudo apt update && sudo apt install build-essential cmake lcov
+      - name: Create conda environment
         run: |
           #export PATH="/usr/share/miniconda/bin:$PATH"
           #alias conda="$CONDA/bin/conda"
           conda create -n zerod python=3.11.4
-          echo "Created environment"
-          strings /usr/share/miniconda/envs/zerod/bin/../lib/libstdc++.so.6 | grep GLIBCXX
-          #echo "test1"
-          #conda install -n zerod libgcc-ng
-          echo "test2"
-          #conda install -n zerod -c conda-forge libgcc-ng
+      - name: Quick-fix dependencies to get correct version numbers
+        if: startsWith(matrix.os, 'ubuntu-latest')
           conda install -n zerod -c conda-forge libstdcxx-ng=13
-          echo "Checked versions"
+      - name: Install svZeroDSolver
           conda run -n zerod pip install -e ".[dev]"
-          strings /usr/share/miniconda/envs/zerod/bin/../lib/libstdc++.so.6 | grep GLIBCXX
       - name: Install Networkx
         run: |
             conda run -n zerod pip install networkx
