+++ .github/workflows/test.yml
@@ -22,10 +22,11 @@ jobs:
           #export PATH="/usr/share/miniconda/bin:$PATH"
           #alias conda="$CONDA/bin/conda"
           conda create -n zerod python=3.11.4
-      - name: Quick-fix dependencies to get correct version numbers
+      - name: Install dependencies to get correct version numbers (Ubuntu)
         #if: startsWith(matrix.os, 'ubuntu-latest')
         if: startsWith(matrix.os, 'ubuntu')
         run: conda install -n zerod -c conda-forge libstdcxx-ng=13 gcc=13
+      - name: Install dependencies to get correct version numbers (MacOS)
         if: startsWith(matrix.os, 'macos')
         run: conda install -n zerod -c conda-forge gcc=13
       - name: Install svZeroDSolver
