+++ .github/workflows/test.yml
@@ -31,8 +31,8 @@ jobs:
       - name: Install dependencies to get correct version numbers (MacOS)
         if: startsWith(matrix.os, 'macos')
         run: |
-          brew install gcc@${{ env.GCC_V }}
-          ln -s /usr/local/bin/gcc-${{ env.GCC_V }} /usr/local/bin/gcc
+          brew install gcc@${GCC_V}
+          ln -s /usr/local/bin/gcc-${GCC_V} /usr/local/bin/gcc
       - name: Install svZeroDSolver
         run: conda run -n zerod pip install -e ".[dev]"
       - name: Install Networkx
