+++ .github/workflows/codechecks.yml
@@ -8,7 +8,7 @@ jobs:
       - uses: actions/checkout@v3
       - name: Install dependencies
         run: |
-          sudo apt install clang-format
+          sudo apt install clang-format-20
       - name: Run clang-format
         run: |
           mkdir Release
