+++ .github/workflows/codechecks.yml
@@ -8,6 +8,7 @@ jobs:
       - uses: actions/checkout@v3
       - name: Install dependencies
         run: |
+          wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
           sudo apt install clang-format-20
       - name: Run clang-format
         run: |
