+++ .github/workflows/codechecks.yml
@@ -8,9 +8,9 @@ jobs:
       - uses: actions/checkout@v3
       - name: Install dependencies
         run: |
-          #wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
-          wget -qO- https://apt.llvm.org/llvm-snapshot.gpg.key | sudo tee /etc/apt/trusted.gpg.d/apt.llvm.org.asc
-          sudo apt-get install clang-format-20
+          wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
+          #wget -qO- https://apt.llvm.org/llvm-snapshot.gpg.key | sudo tee /etc/apt/trusted.gpg.d/apt.llvm.org.asc
+          sudo apt-get install clang-format-19
       - name: Run clang-format
         run: |
           mkdir Release
