### 2025-03-11 32b46d7
Specify GCC and clang-format versions in Github runners (#148)
Co-authored-by: menon-karthik <menon-karthik@users.noreply.github.com>
