+++ paper/paper.md
@@ -11,10 +11,10 @@ authors:
    equal-contrib: true
    affiliation: 1
  - name: Jakob Richter
-   affiliation: 1
+   affiliation: 3
    equal-contrib: true
  - name: Martin R. Pfaller
-   affiliation: 1
+   affiliation: 2
    equal-contrib: true
  - name: Jonathan Pham
    affiliation: 1
@@ -30,9 +30,13 @@ authors:
    affiliation: 1
    corresponding: true
 affiliations:
- - name: Stanford University, Stanford, CA, United States of America
+ - name: Georgia Institute of Technology, Atlanta, GA, United States of America
    index: 1
-date: July 2024
+ - name: Yale University, New Haven, CT, United States of America
+   index: 2
+ - name: Stanford University, Stanford, CA, United States of America
+   index: 3
+date: March 2025
 bibliography: paper.bib
 ---
 
@@ -46,8 +50,8 @@ As part of the [`SimVascular`](https://simvascular.github.io/) open-source proje
 `svZeroDSolver` is written in C++ using a modular object-oriented framework.
 Simply specifying a `.json` dictionary of lumped-parameter "blocks" -- such as blood vessels, junctions between blood vessels, and boundary conditions (along with their associated parameters) -- allows the code to automatically assemble and solve the governing equations corresponding to the user-specified vascular model.
 In addition, the package includes Python and C++ interfaces to facilitate its use with other software packages.
-For example, it can be integrated into Python-based optimization and uncertainty quantification applications [@zanoni2024;@Lee2024;@richter2024bayesian;@menon2024personalizeduncertainty].
-It can also be interfaced with C++/Fortran software for high-fidelity cardiovascular flow simulations, where `svZeroDSolver` can conveniently provide physiological lumped-parameter boundary conditions [@menon2023predictors;@menon2024personalized].
+For example, it can be integrated into Python-based optimization and uncertainty quantification applications [@zanoni2024; @Lee2024; @richter2024bayesian; @menon2024personalizeduncertainty].
+It can also be interfaced with C++/Fortran software for high-fidelity cardiovascular flow simulations, where `svZeroDSolver` can conveniently provide physiological lumped-parameter boundary conditions [@menon2023predictors; @menon2024personalized].
 `svZeroDSolver` includes an application, called `svZeroDCalibrator`, to automatically calibrate parameters of a given zero-dimensional model based independent hemodynamic measurements or high-fidelity simulations -- thus improving the accuracy of zero-dimensional models [@richter2024bayesian].
 It also includes graphical interfaces to interactively create lumped-parameter models for simulations, as well as to visualize the simulated anatomy and hemodynamics.
 
@@ -57,11 +61,11 @@ Non-invasive quantification of patient-specific hemodynamics via computational s
 Computational modeling is also a promising tool for non-invasive and personalized optimization of clinical treatments and surgery [@Marsden2014]. 
 
 Previous work has used several techniques to model cardiovascular blood flow, all of which can be broadly categorized based on their level of fidelity. 
-High-fidelity models generally involve simulations of the full three-dimensional flow-field within anatomical regions of interest [@menon2024cardiovascular;@Updegrove2017]. 
+High-fidelity models generally involve simulations of the full three-dimensional flow-field within anatomical regions of interest [@menon2024cardiovascular; @Updegrove2017]. 
 While these are the most accurate and informative, they are computationally expensive (each simulation can take several hours or days on hundreds of CPU cores) and therefore not practical in typical clinical settings or for applications, such as optimization and uncertainty quantification, which often require thousands of model evaluations.
 On the other end of the spectrum, lumped-parameter or zero-dimensional models provide information about bulk hemodynamics, such as flow rate and pressure, at specific anatomical regions of interest.
[…]
