### 2025-03-13 efec8c2
Update JOSS paper branch with first set of reviewers' comments (#149)
Co-authored-by: menon-karthik <menon-karthik@users.noreply.github.com>
