### 2025-03-13 #7595 @menon-karthik
@philipcardiff I have addressed the comments from the reviewers here: https://github.com/SimVascular/svZeroDSolver/pull/149

Note that we are still trying to figure out [this comment](https://github.com/openjournals/joss-reviews/issues/7595#issuecomment-2614401867) (and [related issue](https://github.com/SimVascular/svZeroDSolver/issues/143)) raised by @marchirschvogel. We are not able to reproduce the issue ourselves. But that might be because we usually use environments to manage dependencies when installing the software (and I am hesitant to try installing without an environment for fear of breaking other things on my system). If using an environment solves the issue for @marchirschvogel  too, we will add it to the docs as the recommended method to install.
