+++ src/model/Block.h
@@ -50,7 +50,7 @@
  * to the global system.
  */
 struct TripletsContributions {
-  TripletsContributions() {};
+  TripletsContributions(){};
   /**
    * @brief Set the number of triplets that the element contributes
    * to the global system.
@@ -58,7 +58,7 @@ struct TripletsContributions {
    * @param E Contributions to E matrix
    * @param D Contributions to dC/dy matrix
    */
-  TripletsContributions(int F, int E, int D) : F(F), E(E), D(D) {};
+  TripletsContributions(int F, int E, int D) : F(F), E(E), D(D){};
   /**
    * @brief Set the number of triplets that the element contributes
    * to the global system.
