+++ .github/workflows/codechecks.yml
@@ -8,7 +8,8 @@ jobs:
       - uses: actions/checkout@v3
       - name: Install dependencies
         run: |
-          sudo apt install clang-format
+          wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
+          sudo apt-get install clang-format-19
       - name: Run clang-format
         run: |
           mkdir Release
