+++ .github/workflows/test.yml
@@ -8,7 +8,10 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-22.04, ubuntu-latest, macos-13, macos-latest]
+        version: [13] # GCC version
       fail-fast: false
+    env:
+      GCC_V: ${{ matrix.version }}
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
@@ -17,12 +20,21 @@ jobs:
       - name: Install ubuntu dependencies
         if: startsWith(matrix.os, 'ubuntu')
         run: sudo apt update && sudo apt install build-essential cmake lcov
-      - name: Install svZeroDSolver
+      - name: Create conda environment
         run: |
           #export PATH="/usr/share/miniconda/bin:$PATH"
           #alias conda="$CONDA/bin/conda"
           conda create -n zerod python=3.11.4
-          conda run -n zerod pip install -e ".[dev]"
+      - name: Install dependencies to get correct version numbers (Ubuntu)
+        if: startsWith(matrix.os, 'ubuntu')
+        run: conda install -n zerod -c conda-forge libstdcxx-ng=${GCC_V} gcc=${GCC_V}
+      - name: Install dependencies to get correct version numbers (MacOS)
+        if: startsWith(matrix.os, 'macos')
+        run: |
+          brew install gcc@${GCC_V}
+          ln -s /usr/local/bin/gcc-${GCC_V} /usr/local/bin/gcc
+      - name: Install svZeroDSolver
+        run: conda run -n zerod pip install -e ".[dev]"
       - name: Install Networkx
         run: |
             conda run -n zerod pip install networkx
