### 2025-03-17 SimVascular/svZeroDSolver#143 @marchirschvogel
I can confirm that installation after recent updates to my system and using the latest code version now works (without a virtual environment).
