### 2025-03-17 #7595 @marchirschvogel
> [@philipcardiff](https://github.com/philipcardiff) I have addressed the comments from the reviewers here: [SimVascular/svZeroDSolver#149](https://github.com/SimVascular/svZeroDSolver/pull/149)
>
> Note that we are still trying to figure out [this comment](https://github.com/openjournals/joss-reviews/issues/7595#issuecomment-2614401867) (and [related issue](https://github.com/SimVascular/svZeroDSolver/issues/143)) raised by [@marchirschvogel](https://github.com/marchirschvogel). We are not able to reproduce the issue ourselves. But that might be because we usually use environments to manage dependencies when installing the software (and I am hesitant to try installing without an environment for fear of breaking other things on my system). If using an environment solves the issue for [@marchirschvogel](https://github.com/marchirschvogel) too, we will add it to the docs as the recommended method to install.

@menon-karthik The issue has now been resolved on my end and I closed it.
