### 2025-03-17 #7595 @marchirschvogel
> I'm missing a "State of the field" section, or at least references to other open source projects that support 0D vascular modeling. Maybe at least some of the contributions that were mentioned in [#7459 (comment)](https://github.com/openjournals/joss-reviews/issues/7459#issuecomment-2470139393) could be discussed in terms of commonalities/differences to svZeroDSolver.

I've finished the review checklist now. I can confirm that both ways to install the software (via pip and from source) work on my end, and that the functionality of the software is given and the results from the test case documented can be reproduced.
Above-mentioned "State of the field" section would be a last missing point of the manuscript for me to recommend publication in JOSS.
