### 2025-03-17 #7595 @menon-karthik
@marchirschvogel Thanks for the update! I didn't specify in my previous comment that I did add lines 100-115 in the paper to address the "state of the field" comment. I haven't made it a separate section, but I have added a discussion of similar repositories and what makes this package unique. I can make it a separate section if required too.
