### 2025-03-18 #7595 @marchirschvogel
> [@marchirschvogel](https://github.com/marchirschvogel) Thanks for the update! I didn't specify in my previous comment that I did add lines 100-115 in the paper to address the "state of the field" comment. I haven't made it a separate section, but I have added a discussion of similar repositories and what makes this package unique. I can make it a separate section if required too.

OK, thanks. This looks good to me. I now can recommend publication in JOSS.
