### 2025-03-20 ba2e1c1
simple check if factorization was successful
