+++ src/algebra/SparseSystem.cpp
@@ -100,6 +100,8 @@ void SparseSystem::update_jacobian(double time_coeff_ydot,
 
 void SparseSystem::solve() {
   solver->factorize(jacobian);
+  if (solver->info() != Eigen::Success)
+    throw std::runtime_error("System is singular");
   dydot.setZero();
   dydot += solver->solve(residual);
 }
