### 2025-03-24 #7595 @TS-CUBED
Hello, apologies for the delay in my review. I have now gone through all my notes and checklist.

1. One outstanding thing I have is again related  to the license (and since this is JOSS, I am picky): The phrase "All rights reserved" is considered as incompatible with the MIT license, as far as I know. So it would be best, to simply adopt the MIT license directly, rather than attempting to write one's own.

2. With only one example from 2018 the comparison to other packages could be more extensive.

Other than that I can recommend publication.
