### 2025-03-25 9e33caa
fix resistance and inductance test. capacitance still unsure
