### 2025-03-26 7de6967
fix capacitance test case
