### 2025-03-26 4325011
remove untested calibrator code
