### 2025-03-31 #7595 @menon-karthik
@TS-CUBED thanks for your comments!

Your point about the MIT license makes sense. Since svZeroDSolver is part of the broader SimVascular project, I will need to confer with the developers of the other repositories too about the license since we ideally like to have a uniform license for all repositories under the project.

I will also do a more extensive literature search for the state of the field and update the paper.
