### 2025-03-31 86d74d3
added refs for 1d and 0d packages
