+++ paper/paper.bib
@@ -329,3 +329,80 @@ author = {Pasquale Claudio Africa and Ivan Fumagalli and Michele Bucelli and Alb
 @article{Hirschvogel2024, doi = {10.21105/joss.05744}, url = {https://doi.org/10.21105/joss.05744}, year = {2024}, publisher = {The Open Journal}, volume = {9}, number = {93}, pages = {5744}, author = {Marc Hirschvogel}, title = {Ambit – A FEniCS-based cardiovascular multi-physics solver}, journal = {Journal of Open Source Software} }
 
 @article{Agdestein2018, doi = {10.21105/joss.01107}, url = {https://doi.org/10.21105/joss.01107}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {32}, pages = {1107}, author = {Syver D. Agdestein and Kristian Valen-Sendstad and Alexandra K. Diem}, title = {Artery.FE: An implementation of the 1D blood flow equations in FEniCS}, journal = {Journal of Open Source Software} }
+
+@article{Diem-2017,
+ author = {Diem, Alexandra K. and Bressloff, Neil W.},
+ doi = {10.5334/jors.159},
+ journal = {Journal of Open Research Software},
+ keyword = {en_US},
+ month = {Jun},
+ title = {VaMpy: A Python Package to Solve 1D Blood Flow Problems},
+ year = {2017}
+}
+
+@article{REGAZZONI2021104641,
+title = {Accelerating the convergence to a limit cycle in 3D cardiac electromechanical simulations through a data-driven 0D emulator},
+journal = {Computers in Biology and Medicine},
+volume = {135},
+pages = {104641},
+year = {2021},
+issn = {0010-4825},
+doi = {https://doi.org/10.1016/j.compbiomed.2021.104641},
+url = {https://www.sciencedirect.com/science/article/pii/S0010482521004352},
+author = {F. Regazzoni and A. Quarteroni},
+keywords = {Mathematical modeling, Cardiac modeling, Cardiac electromechanics, Multiphysics models, Lumped-parameter models}
+}
+
+@article{benemerito2024openbf,
+  title={openBF: an open-source finite volume 1D blood flow solver},
+  author={Benemerito, I and Melis, A and Wehenkel, Antoine and Marzo, A},
+  journal={Physiological Measurement},
+  volume={45},
+  number={12},
+  pages={125002},
+  year={2024},
+  publisher={IOP Publishing}
+}
+
+@article{cellml,
+url = {https://doi.org/10.1515/jib-2020-0021},
+title = {CellML 2.0},
+author = {Michael Clerx and Michael T. Cooling and Jonathan Cooper and Alan Garny and Keri Moyle and David P. Nickerson and Poul M. F. Nielsen and Hugh Sorby},
+pages = {20200021},
+volume = {17},
+number = {2-3},
+journal = {Journal of Integrative Bioinformatics},
+doi = {doi:10.1515/jib-2020-0021},
+year = {2020}
+}
+
+@inproceedings{alastruey2012arterial,
+  title={Arterial pulse wave haemodynamics},
+  author={Alastruey, Jordi and Parker, Kim H and Sherwin, Spencer J},
+  booktitle={11th international conference on pressure surges},
+  pages={401--443},
+  year={2012},
+  organization={Virtual PiE Led t/a BHR Group}
+}
+
+@article{heldt2010cvsim,
+  title={CVSim: an open-source cardiovascular simulator for teaching and research},
+  author={Heldt, Thomas and Mukkamala, Ramakrishna and Moody, George B and Mark, Roger G},
+  journal={The open pacing, electrophysiology \& therapy journal},
+  volume={3},
+  pages={45},
+  year={2010}
+}
+
+@article{HUTTARY2017104,
[…]
