+++ paper/paper.md
@@ -92,9 +92,17 @@ This combination of features makes `svZeroDSolver` uniquely applicable to a wide
 While there are other open-source projects that provide the functionality for cardiovascular flow modeling, and specifically zero-dimensional flow modeling, `svZeroDSolver` has several features that distinguish it from previous work. 
 For one, prior packages have primarily focused on multi-physics finite element modeling for cardiovascular biomechanics [@crimson2021; @Zhu2022; @AFRICA2024109039; @Hirschvogel2024]. 
 While these projects allow the implementation of simple zero-dimensional models, usually as boundary conditions to three-dimensional models, the primary focus is on the modeling of full three-dimensional fluid and tissue mechanics. 
-There are also packages aimed at reduced-order modeling for cardiovascular flows [@Agdestein2018]. In particular, the `SimVascular` project includes `svOneDSolver` for the purpose of one-dimensional blood flow modeling.
-In contrast to these packages, the purpose of `svZeroDSolver` is to provide a modular framework specifically for simulating zero-dimensional flows in a variety of simple and complex anatomies -- by leveraging the modular nature of the code.
-The unique features listed above allow the use of `svZeroDSolver` both as a stand-along zero-dimensional flow solver for unique and patient-specific anatomies, as well as in conjunction with the aforementioned multi-physics solvers as boundary conditions, for parameters estimation and uncertainty quantification, or even as an instructional tool using its graphical interfaces.  
+There are also packages aimed at reduced-order modeling for cardiovascular flows.
+In particular, the `SimVascular` project includes `svOneDSolver` for the purpose of one-dimensional blood flow modeling.
+Another popular package for one-dimensional blood flow simulations is `Nektar1D` [@alastruey2012arterial].
+Similarly, the `Artery.FE` package implements one-dimensional blood flow modeling using the FEniCS finite element framework [@Agdestein2018], the `VaMpy` toolkit includes a package for modeling one-dimensional blood flow using the Lax-Wendroff finite difference method [@Diem-2017], and `openBF` is a finite volume implementation of one-dimensional blood flow [@benemerito2024openbf].
+In the zero-dimensional modeling context, `CRIMSON` [@crimson2021] and `lifex-cfd` [@AFRICA2024109039] include the ability to simulate simple zero-dimensional blood flow models, primarily as boundary conditions to three-dimensional simulations.
+However, their focus is on multi-physics simulations of cardiovascular biomechanics, therefore they are not stand-alone and modular zero-dimensional flow solvers.
+The `CellML` and `CVSim`  packages include a limited set of stand-alone zero-dimensional flow models for specific anatomies/applications [@cellml; @heldt2010cvsim], but they do not provide the modular functionality to specify unique anatomical models.
+In addition, there have been other packages that use zero-dimensional modeling techniques with a focus on statistical analysis and cardiac electromechanics [@HUTTARY2017104; @REGAZZONI2021104641].
+However, these packages are either not focused on zero-dimensional modeling or use MATLAB implementations, which require software licenses and are not free to use. 
+In contrast, the purpose of `svZeroDSolver` is to provide an open-source framework specifically for simulating zero-dimensional flows in a variety of simple and complex anatomies that can be designed in a user-specific and application-specific manner -- by leveraging the modular nature of the code.
+The unique features listed above allow the use of `svZeroDSolver` both as a stand-alone zero-dimensional flow solver for unique and patient-specific anatomies, as well as in conjunction with the aforementioned multi-physics solvers as boundary conditions, for parameters estimation and uncertainty quantification, or even as an instructional tool using its graphical interfaces.  
 
 ![Various zero-dimensional "blocks" included in `svZeroDSolver` at the time of writing.\label{fig:blocks}](blocks.png){ width=80% }
 
