+++ paper/paper.bib
@@ -406,3 +406,14 @@ url = {https://www.sciencedirect.com/science/article/pii/S0010482517301439},
 author = {Rudolf Huttary and Leonid Goubergrits and Christof Schütte and Stefan Bernhard},
 keywords = {Windkessel elements, Lumped models, 0D modeling, Multi-compartment modeling, Cardiovascular simulation, Distributed parameter modeling, Clinical data set, Coarctation of aorta, Patient-specific models, Disease-specific models, Multiscale modeling}
 }
+
+@article{rosalia2021object,
+  title={Object-oriented lumped-parameter modeling of the cardiovascular system for physiological and pathophysiological conditions},
+  author={Rosalia, Luca and Ozturk, Caglar and Van Story, David and Horvath, Markus A and Roche, Ellen T},
+  journal={Advanced Theory and Simulations},
+  volume={4},
+  number={3},
+  pages={2000216},
+  year={2021},
+  publisher={Wiley Online Library}
+}
