### 2025-04-02 75422ee
compare dataframes more efficiently, recompute some reference solutions
