+++ tests/utils.py
@@ -2,13 +2,15 @@ import json
 import os
 import subprocess
 from tempfile import TemporaryDirectory
+import matplotlib.pyplot as plt
+import seaborn as sns
 
 import numpy as np
 import pandas as pd
 
 # global boolean to perform coverage testing
 # (run executables instead of Python interface, much slower)
-from pytest import coverage
+import coverage
 
 import pysvzerod
 
@@ -58,61 +60,47 @@ def run_with_reference(
         ):
 
 
-    res, config = execute_pysvzerod(test_config, "solver")
+    res = pysvzerod.simulate(test_config)
 
-    if res.shape[1] == 6:
-        # we have a result with fields [name, time, p_in, p_out, q_in, q_out]
-        for field in ["pressure_in", "pressure_out", "flow_in", "flow_out"]:
-            if "pressure" in field:
-                assert np.isclose(res[field].to_numpy().all(), ref[field].to_numpy().all(), rtol=RTOL_PRES)
-            elif "flow" in field:
-                assert np.isclose(res[field].to_numpy().all(), ref[field].to_numpy().all(), rtol=RTOL_FLOW)
+    if res.shape[1] >= 6:
+        # we have a result with fields [name, time, p_in, p_out, q_in, q_out] SOME HAVE GREATTER LENGTH NEED TO ADDRESS
+        pressure_cols = ["pressures_in", "pressure_out"]
+        flow_cols = ["flow_in", "flow_out"]
+
+        bool_df = pd.DataFrame(index=res.index, columns=res.columns, dtype=bool)
+
+        for col in res.columns:
+            tol = RTOL_PRES if col in pressure_cols else RTOL_FLOW if col in flow_cols else None
+            if tol is not None:
+                check = np.abs(res[col] - ref[col]) < tol
+                bool_df[col] = check  # True if difference is below tolerance
+
+        if not bool_df.all().all():  # Check if any value exceeded tolerance
+            # Extract only differing rows/columns for a cleaner error message
+            differing_locs = np.where(~bool_df)  # Shows only the out-of-tolerance values
+            differing_indices = res.index[differing_locs[0]]
+            differing_columns = res.columns[differing_locs[1]]
+
+            if differing_indices.any():  # If any differences are found
+                diff_locations = list(zip(differing_indices, differing_columns))
+                raise AssertionError(f"Differences exceed tolerance at:\n{diff_locations}")
+
     else:
         # we have a result with fields [name, time, y] and the result must be compared based on the name field. name is of format [flow:vessel:outlet]
         # we will compare the average of each branch
-        avg_res_flow = []
-        avg_ref_flow = []
-        avg_res_pres = []
-        avg_ref_pres = []
-        for index, row in res.iterrows():
 
-            if "flow" in row["name"]:
-                if row["name"] == res.iloc[index + 1]["name"]:
-                    # we are compilng the results for a branch
-                    avg_ref_flow.append(ref.loc[row.name].y)
-                    avg_res_flow.append(row.y)
-                elif avg_res_flow == []:
-                    # there is only one result for this branch
-                    assert np.isclose(row.y, ref.loc[row.name].y, rtol=RTOL_FLOW)
-                else:
-                    # we are on the last result for this branch
-                    avg_ref_flow.append(ref.loc[row.name].y)
-                    avg_res_flow.append(row.y)
-                    assert np.isclose(np.array(avg_res_flow).all(), np.array(avg_ref_flow).all(), rtol=RTOL_FLOW)
-                    avg_res_flow = []
-                    avg_ref_flow = []
-                    
[…]
