### 2025-04-02 a56a131
Update JOSS paper references and license (#163)
Co-authored-by: menon-karthik <menon-karthik@users.noreply.github.com>
Co-authored-by: Martin Pfaller <pfaller@stanford.edu>
Co-authored-by: Nick Dorn <ndorn@stanford.edu>
