### 2025-04-02 35c2260
Update JOSS paper references and license (#163)
Co-authored-by: menon-karthik <menon-karthik@users.noreply.github.com>
Co-authored-by: Martin Pfaller <pfaller@stanford.edu>
Co-authored-by: Nick Dorn <ndorn@stanford.edu>
