### 2025-04-02 df63bd9
Clean up code and update ref vs res plotting
