### 2025-04-02 5d12245
Remove unused plotting functions from utils.py
