+++ tests/utils.py
@@ -2,8 +2,6 @@ import json
 import os
 import subprocess
 from tempfile import TemporaryDirectory
-import matplotlib.pyplot as plt
-import seaborn as sns
 
 import numpy as np
 import pandas as pd
@@ -146,103 +144,3 @@ def get_result(result_array, field, branch, time_step):
     # extract result
     return result_array[field][branch][time_step]
 
-
-def plot_res_vs_ref_sol(testfile):
-    '''
-    plot the results of a test case against the reference solution
-    :param testfile: path to the test case json file
-    '''
-
-    # read reference solution
-    ref_sol = pd.read_json(os.path.join(this_file_dir, "cases", "results", "result_" + os.path.basename(testfile)))
-    # read result
-    result = pysvzerod.simulate(testfile)
-
-    ref_marker = "--"  # marker for the reference solution
-    
-    if result.shape[1] == 6:
-        # we have a result with fields [name, time, p_in, p_out, q_in, q_out]
-        # plot results
-        fig, ax = plt.subplots(2, 2, figsize=(10, 10))
-        # plot pressure_in
-        ax[0, 0].plot(result["time"], result["pressure_in"], label="pysvzerod")
-        ax[0, 0].plot(ref_sol["time"], ref_sol["pressure_in"], label="reference solution", linestyle=ref_marker)
-        ax[0, 0].set_title("Pressure In")
-        ax[0, 0].set_xlabel("Time")
-        ax[0, 0].set_ylabel("Pressure [Pa]")
-        ax[0, 0].legend()
-        # plot pressure_out
-        ax[0, 1].plot(result["time"], result["pressure_out"], label="pysvzerod")
-        ax[0, 1].plot(ref_sol["time"], ref_sol["pressure_out"], label="reference solution", linestyle=ref_marker)
-        ax[0, 1].set_title("Pressure Out")
-        ax[0, 1].set_xlabel("Time")
-        ax[0, 1].set_ylabel("Pressure [Pa]")
-        ax[0, 1].legend()
-        # plot flow_in
-        ax[1, 0].plot(result["time"], result["flow_in"], label="pysvzerod")
-        ax[1, 0].plot(ref_sol["time"], ref_sol["flow_in"], label="reference solution", linestyle=ref_marker)
-        ax[1, 0].set_title("Flow In")
-        ax[1, 0].set_xlabel("Time")
-        ax[1, 0].set_ylabel("Flow [m^3/s]")
-        ax[1, 0].legend()
-        # plot flow_out
-        ax[1, 1].plot(result["time"], result["flow_out"], label="pysvzerod")
-        ax[1, 1].plot(ref_sol["time"], ref_sol["flow_out"], label="reference solution", linestyle=ref_marker)
-        ax[1, 1].set_title("Flow Out")
-        ax[1, 1].set_xlabel("Time")
-        ax[1, 1].set_ylabel("Flow [m^3/s]")
-        ax[1, 1].legend()
-        # show plot
-        plt.tight_layout()
-        plt.show()
-
-    else:
-        # we have a result with fields [name, time, y] and the result must be compared based on the name field. name is of format [flow:vessel:outlet]
-
-        
-        grouped_by_name = result.groupby("name")
-
-        # plot the groups with flow in the name using seaborn lineplot
-        fig, ax = plt.subplots(1, 2, figsize=(10, 5))
-        for name, group in grouped_by_name:
-            if "flow" in name:
-                group = group.sort_values("time")
-                group["y"] = group["y"].astype(float)
-                group["time"] = group["time"].astype(float)
-                # plot the group in a subplot
-                ax[0].plot(group["time"], group["y"], label=name)
-            elif "pressure" in name:
-                group = group.sort_values("time")
-                group["y"] = group["y"].astype(float)
[…]
