### 2025-04-02 669be6a
some values close to zero still have relative difference above tolerance
