### 2025-04-03 23bebd7
Refine relative difference calculation and tolerance check in run_with_reference function
