### 2025-04-03 a78a67b
Fix relative difference calculation to handle absolute values and remove unused main execution block
