### 2025-04-03 b1cc0e7
Replace coverage import with pytest coverage for improved testing integration
