### 2025-04-03 dc4b1a4
Simplify error message for tolerance assertion in run_with_reference function
