### 2025-04-03 c1a742b
Move license to LICENSE.md and update file headers, fixes #164
