### 2025-04-03 58e2c9b
update solver flow tolerance to see if tests pass on Ubuntu
