### 2025-04-03 aeb1567
Move license to LICENSE.md and update file headers, fixes #164
