+++ applications/svZeroDVisualization/dirgraph_connections.py
@@ -1,10 +1,7 @@
 # coding=utf-8
 
-# This file is part of svZeroDSolver licensed under Stanford University, The Regents of the University of 
-#                                                   California, and others.
-# 
+# This file is part of svZeroDSolver licensed under the MIT License
 # See the LICENSE.md file for license information
-
 
 import numpy as np
 from dirgraph_wire import wire
