### 2025-04-03 d9ab50a
Update solver + test tolerance values and enhance reference solution computation functionality
