+++ src/model/Model.cpp
@@ -60,9 +60,6 @@ Model::Model() {
       {"RegazzoniValve", block_factory<RegazzoniValve>()},
       {"RegazzoniChamber", block_factory<RegazzoniChamber>()},
       {"KungVentricle", block_factory<RegazzoniChamber>()}};
-
-
-      
 }
 
 Model::~Model() {}
@@ -210,7 +207,7 @@ void Model::finalize() {
   }
 
   if (cardiac_cycle_period < 0.0) {
-    cardiac_cycle_period =  1.0;
+    cardiac_cycle_period = 1.0;
   }
 }
 
