+++ src/solve/SimulationParameters.h
@@ -53,8 +53,8 @@ struct SimulationParameters {
   double sim_time_step_size{0.0};   ///< Simulation time step size
   double sim_abs_tol{0.0};          ///< Absolute tolerance for simulation
   double sim_cardiac_period{-1.0};  ///< Cardiac period
-  int sim_num_cycles{0};     ///< Number of cardiac cycles to simulate
-  int sim_pts_per_cycle{0};  ///< Number of time steps per cardiac cycle
+  int sim_num_cycles{0};            ///< Number of cardiac cycles to simulate
+  int sim_pts_per_cycle{0};         ///< Number of time steps per cardiac cycle
   bool use_cycle_to_cycle_error{
       false};  ///< If model does not have RCR boundary conditions, simulate
                ///< model to convergence (based on cycle-to-cycle error of last
