+++ tests/test_dirgraph.py
@@ -22,7 +22,8 @@ excluded_files = [
     'coupledBlock_closedLoopHeart_withCoronaries.json',
     'closedLoopHeart_singleVessel_mistmatchPeriod.json',
     'pulsatileFlow_R_RCR_mismatchPeriod.json',
-    'pulsatileFlow_CStenosis_steadyPressure_definedPeriod.json'
+    'pulsatileFlow_CStenosis_steadyPressure_definedPeriod.json',
+    'double_pulsatileFlow_CRL.json'
 ]
 
 # Generate the list of JSON files to test
