### 2025-04-08 de92134
Model definition clean up
