+++ src/model/Model.cpp
@@ -205,10 +205,6 @@ void Model::finalize() {
   for (auto &block : blocks) {
     block->setup_model_dependent_params();
   }
-
-  if (cardiac_cycle_period < 0.0) {
-    cardiac_cycle_period = 1.0;
-  }
 }
 
 int Model::get_num_blocks(bool internal) const {
