+++ src/model/BloodVesselNew.cpp
@@ -36,25 +36,25 @@ void BloodVesselNew::setup_dofs(DOFHandler &dofhandler) {
 
 void BloodVesselNew::update_constant(SparseSystem &system,
                                   std::vector<double> &parameters) {
-  // Get parameters
-  double capacitance = parameters[global_param_ids[ParamId::CAPACITANCE]];
-  double inductance = parameters[global_param_ids[ParamId::INDUCTANCE]];
-  double resistance = parameters[global_param_ids[ParamId::RESISTANCE]];
+  // Hard code constant parameters
+
+  const double rho = 1000.0000; 
+  const double d = 0.0142; 
+  const double Ro = 0.0236; 
+  const double W1 = 1.0000; 
+  const double W2 = 1.0000; 
+  const double eta = 70.0000; 
+  const double a = 1.0000; 
+  const double sigma_o = 124000.0000; 
 
   // Set element contributions
   // coeffRef args are the indices (i,j) of the matrix
   // global_eqn_ids: number of rows in the matrix, set in setup_dofs
   // global_var_ids: number of columns, organized as pressure and flow of all
   // inlets and then all outlets of the block
-  system.E.coeffRef(global_eqn_ids[0], global_var_ids[3]) = -inductance;
-  system.E.coeffRef(global_eqn_ids[1], global_var_ids[0]) = -capacitance;
-  system.E.coeffRef(global_eqn_ids[1], global_var_ids[1]) =
-      capacitance * resistance;
-  system.F.coeffRef(global_eqn_ids[0], global_var_ids[0]) = 1.0;
-  system.F.coeffRef(global_eqn_ids[0], global_var_ids[1]) = -resistance;
-  system.F.coeffRef(global_eqn_ids[0], global_var_ids[2]) = -1.0;
-  system.F.coeffRef(global_eqn_ids[1], global_var_ids[1]) = 1.0;
-  system.F.coeffRef(global_eqn_ids[1], global_var_ids[3]) = -1.0;
+ 
+// deleted E and F matrices %%%%%%
+
 }
 
 void BloodVesselNew::update_solution(
@@ -62,26 +62,78 @@ void BloodVesselNew::update_solution(
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &y,
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &dy) {
   // Get parameters
-  double capacitance = parameters[global_param_ids[ParamId::CAPACITANCE]];
-  double stenosis_coeff =
-      parameters[global_param_ids[ParamId::STENOSIS_COEFFICIENT]];
-  double q_in = y[global_var_ids[1]];
-  double dq_in = dy[global_var_ids[1]];
-  double stenosis_resistance = stenosis_coeff * fabs(q_in);
+  double rho = parameters[global_param_ids[ParamId::rho]];
+  double d = parameters[global_param_ids[ParamId::d]];
+  double Ro = parameters[global_param_ids[ParamId::Ro]];
+  double W1 = parameters[global_param_ids[ParamId::W1]];
+  double W2 = parameters[global_param_ids[ParamId::W2]];
[…]
