+++ src/model/BloodVesselNew.cpp
@@ -31,31 +31,67 @@
 #include "BloodVesselNew.h"
 
 void BloodVesselNew::setup_dofs(DOFHandler &dofhandler) {
-  Block::setup_dofs_(dofhandler, 2, {});
+  Block::setup_dofs_(dofhandler, 8, {"r","v","S","k","tau","V"});
 }
 
 void BloodVesselNew::update_constant(SparseSystem &system,
                                   std::vector<double> &parameters) {
   // Hard code constant parameters
 
-  const double rho = 1000.0000; 
-  const double d = 0.0142; 
-  const double Ro = 0.0236; 
-  const double W1 = 1.0000; 
-  const double W2 = 1.0000; 
-  const double eta = 70.0000; 
-  const double a = 1.0000; 
-  const double sigma_o = 124000.0000; 
+  // const double rho = 1000.0; 
+  // const double d = 0.0142; 
+  // const double Ro = 0.0236; 
+  // const double W1 = 1.0000; 
+  // const double W2 = 1.0000; 
+  // const double eta = 70.0000; 
+  // const double a = 1.0000; 
+  // const double sigma_o = 124000.0000; 
 
-  // Set element contributions
+
+   // Set element contributions
   // coeffRef args are the indices (i,j) of the matrix
   // global_eqn_ids: number of rows in the matrix, set in setup_dofs
   // global_var_ids: number of columns, organized as pressure and flow of all
   // inlets and then all outlets of the block
  
 // deleted E and F matrices %%%%%%
-
 }
+
+// void ChamberElastanceInductor::update_time(SparseSystem &system,
+//   std::vector<double> &parameters) {
+// get_elastance_values(parameters);
+
+// system.C.coeffRef(global_eqn_ids[2]) = tau_ - a * sigma_o + a * tau;
+// system.dC_dy.coeffRef(global_eqn_ids[2], global_var_ids[8]) = a;
+// }
+
+// void ChamberElastanceInductor::get_elastance_values(
+// std::vector<double> &parameters) {
+// double Emax = parameters[global_param_ids[ParamId::EMAX]];
+// double Emin = parameters[global_param_ids[ParamId::EMIN]];
+// double Vrd = parameters[global_param_ids[ParamId::VRD]];
+// double Vrs = parameters[global_param_ids[ParamId::VRS]];
+// double t_active = parameters[global_param_ids[ParamId::TACTIVE]];
+// double t_twitch = parameters[global_param_ids[ParamId::TTWITCH]];
+
+// auto T_cardiac = model->cardiac_cycle_period;
+// auto t_in_cycle = fmod(model->time, T_cardiac);
+
+// double t_contract = 0;
+// if (t_in_cycle >= t_active) {
+// t_contract = t_in_cycle - t_active;
+// }
+
+// double act = 0;
+// if (t_contract <= t_twitch) {
+// act = -0.5 * cos(2 * M_PI * t_contract / t_twitch) + 0.5;
+// }
+
+// Vrest = (1.0 - act) * (Vrd - Vrs) + Vrs;
+// a = (Emax - Emin) * act + Emin; // CHANGED "ELAS" TO "a" TO MATCH REST OF SCRIPT
+
+// }
+
 
[…]
