### 2025-04-17 #7595 @marchirschvogel
Two minor things:
- l. 32: I think "model based independent hemodynamic measurements" should be "model based **on** independent ..."
- Regarding the reference to solver Ambit (Hirschvogel 2024): The solver actually can be used as stand-alone 0D solver, too (not only 3D-0D). However, it surely does not provide the degree of modularity that svZeroD does. Maybe this can be put into context slightly differently.
Thanks.
