### 2025-04-17 38911ef
adjust solver and testing tolerances such that tests pass, and add testing guide README
