### 2025-04-23 e08d669
updated active function
