+++ src/algebra/SparseSystem.cpp
@@ -100,6 +100,7 @@ void SparseSystem::update_jacobian(double time_coeff_ydot,
 
 void SparseSystem::solve() {
   solver->factorize(jacobian);
+  // std::cout << Eigen::MatrixXd(jacobian) << std::endl;
   if (solver->info() != Eigen::Success) {
     throw std::runtime_error(
         "System is singular. Check your model (connections, boundary "
