### 2025-04-24 8843b45
adding "<algorithm.h>" library to header and updating math library to "<cmath.h>"
