### 2025-04-24 f218e2a
adding macros for windows interface to emulate <dlfcn.h> library since the dynamic load library is only supported for POSIX systems
