### 2025-04-24 6660bab
add WIN32 specific cmake changes to correctly compile libraries with mingw compiler
