### 2025-04-24 486fc00
adding windows build, testing, and packaging to test workflow
