+++ .github/workflows/test.yml
@@ -7,48 +7,86 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-22.04, ubuntu-latest, macos-13, macos-latest]
+        os: [ubuntu-22.04, ubuntu-latest, macos-13, macos-latest, windows-latest]
         version: [13] # GCC version
       fail-fast: false
     env:
       GCC_V: ${{ matrix.version }}
+      CONDA_ENV: zerod
+      PYTHONPATH: ${{ github.workspace }}
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
         with:
           auto-update-conda: true
+          activate-environment: ${{env.CONDA_ENV}}
+          python-version: "3.11.4"
       - name: Install ubuntu dependencies
         if: startsWith(matrix.os, 'ubuntu')
         run: sudo apt update && sudo apt install build-essential cmake lcov
-      - name: Create conda environment
-        run: |
-          #export PATH="/usr/share/miniconda/bin:$PATH"
-          #alias conda="$CONDA/bin/conda"
-          conda create -n zerod python=3.11.4
+
       - name: Install dependencies to get correct version numbers (Ubuntu)
         if: startsWith(matrix.os, 'ubuntu')
-        run: conda install -n zerod -c conda-forge libstdcxx-ng=${GCC_V} gcc=${GCC_V}
+        run: conda install -c conda-forge libstdcxx-ng=${GCC_V} gcc=${GCC_V}
+
       - name: Install dependencies to get correct version numbers (MacOS)
         if: startsWith(matrix.os, 'macos')
         run: |
           brew install gcc@${GCC_V}
           ln -s /usr/local/bin/gcc-${GCC_V} /usr/local/bin/gcc
-      - name: Install svZeroDSolver
-        run: conda run -n zerod pip install -e ".[dev]"
+
+      - name: Install dependencies for windows
+        if: startsWith(matrix.os, 'windows')
+        shell: pwsh
+        run: |
+          choco install mingw --no-progress
+          conda install -y -c conda-forge cmake graphviz python-graphviz pydot
+          $Env:CMAKE_GENERATOR = 'MinGW Makefiles'
+          Write-Host "→ Using CMAKE_GENERATOR = $Env:CMAKE_GENERATOR"
+          pip install --upgrade cmake-setuptools
+
+      - name: Install POISX-like svZeroDSolver
+        if: ${{!startsWith(matrix.os, 'windows')}}
+        run: conda run pip install -e ".[dev]"
+
+      - name: Install Windows svZeroDSolver
+        if: startsWith(matrix.os, 'windows')
+        shell: pwsh
+        run: |
+          pip install --no-build-isolation -v .[dev]
+          pip show pysvzerod
+
       - name: Install Networkx
         run: |
-            conda run -n zerod pip install networkx
+            conda run pip install networkx
+
       - name: Test the build
         run: |
           cd tests
-          conda run -n zerod pytest -v --durations=0 --ignore=test_dirgraph.py
-      - name: Build using CMake
+          conda run pytest -v --durations=0 --ignore=test_dirgraph.py
+
+      - name: Build using CMake for POISX-like Systems
+        if: ${{!startsWith(matrix.os, 'windows')}}
         run: |
           mkdir Release
           cd Release
           cmake -DCMAKE_BUILD_TYPE=Release -DENABLE_DISTRIBUTION=ON ..
           make -j2
-      - name: Test interface
+
+      - name: Build using CMake for Windows Systems
+        if: startsWith(matrix.os, 'windows')
+        shell: pwsh
+        run: |
+          mkdir Release
+          cd Release
+          cmake -G "MinGW Makefiles" `
[…]
