### 2025-04-24 2205244
ensure that mingw proceeds pip installation
