### 2025-04-24 be180d1
revert to <math.h> library for poisx systems
