### 2025-04-24 43ddaa2
add standard target location across poisx and windows
