+++ CMakeLists.txt
@@ -280,6 +280,10 @@ add_custom_target(codeformat
 add_custom_target(codecheck
   COMMAND find ${SDIR}/*.h ${SDIR}/*.cpp | xargs clang-format -style=file:${PROJECT_SOURCE_DIR}/.clang-format --dry-run --Werror)
 
+set_target_properties(pysvzerod PROPERTIES
+        LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/python
+)
+
 # standard install target locations
 install(TARGETS pysvzerod
         LIBRARY DESTINATION "${CMAKE_INSTALL_LIBDIR}"
