### 2025-04-24 114105a
satisfy modern clang format with <cmath> header instead of <math.h>
