### 2025-04-24 21d0999
remove blank line to satisfy the clang-formatting-19 style
