+++ src/model/Parameter.h
@@ -35,7 +35,6 @@
 #define SVZERODSOLVER_MODEL_PARAMETER_HPP_
 
 #include <cmath>
-
 #include <iostream>
 #include <numeric>
 #include <vector>
