+++ src/model/Parameter.h
@@ -34,8 +34,8 @@
 #ifndef SVZERODSOLVER_MODEL_PARAMETER_HPP_
 #define SVZERODSOLVER_MODEL_PARAMETER_HPP_
 
-#include <math.h>
-
+#include <algorithm>
+#include <cmath>
 #include <iostream>
 #include <numeric>
 #include <vector>
