### 2025-04-25 #7595 @TS-CUBED
All good from my side.

But I would be interested where @menon-karthik sees in [CirculatorySystemModels.jl](https://github.com/TS-CUBED/CirculatorySystemModels.jl) in the bigger picture.

Note: not fishing for citations here. Happy for the paper to be published as is.
