### 2025-04-28 #7595 @philipcardiff
@menon-karthik : FYI, https://github.com/SimVascular/svZeroDSolver/pull/173.

Also, work through the bullet points above.
