### 2025-04-28 cdfcde5
use pressure difference for capacitance
