+++ src/model/BloodVesselNew.cpp
@@ -41,7 +41,10 @@ void BloodVesselNew::update_solution(
     SparseSystem &system, std::vector<double> &parameters,
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &y,
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &dy) {
+    
+  get_elastance_values(parameters);  
   // Get parameters
+  
   double rho = parameters[global_param_ids[ParamId::rho]];
   double d = parameters[global_param_ids[ParamId::d]];
   double Ro = parameters[global_param_ids[ParamId::Ro]];
@@ -74,7 +77,8 @@ void BloodVesselNew::update_solution(
 
   // Set element contributions
   system.C(global_eqn_ids[0]) = -Pout*pow(r/Ro+1.0,2.0)+d*rho*v_+(S*d*(r/Ro+1.0))/Ro;
-  system.C(global_eqn_ids[1]) = -S+tau-(1.0/pow(r/Ro+1.0,6.0)*4.0-4.0)*(W1+W2*pow(r/Ro+1.0,2.0))+1.0/(Ro*Ro)*eta*(1.0/pow(r/Ro+1.0,1.2E+1)*2.0-1.0)*(Ro*2.0+r*2.0)*2.0;
+  system.C(global_eqn_ids[1]) = -S+tau-(1.0/pow(r/Ro+1.0,6.0)*4.0-4.0)*(W1+W2*pow(r/Ro+1.0,2.0))-(eta*r_*(1.0/pow(r/Ro+1.0,1.2E+1)*2.0-1.0)*((r*2.0)/Ro+2.0)*2.0)/Ro;
+  system.C(global_eqn_ids[2]) = tau_-a_plus*sigma_o+a*tau;
   system.C(global_eqn_ids[3]) = Qin-Qout-V_;
   system.C(global_eqn_ids[4]) = -V_+(Ro*Ro)*v*M_PI*pow(r/Ro+1.0,2.0)*4.0;
   system.C(global_eqn_ids[5]) = r_-v;
@@ -83,9 +87,10 @@ void BloodVesselNew::update_solution(
   system.dC_dy.coeffRef(global_eqn_ids[0], global_var_ids[2]) = -pow(r/Ro+1.0,2.0);
   system.dC_dy.coeffRef(global_eqn_ids[0], global_var_ids[4]) = -1.0/(Ro*Ro)*(Pout*Ro*2.0-S*d+Pout*r*2.0);
   system.dC_dy.coeffRef(global_eqn_ids[0], global_var_ids[6]) = 1.0/(Ro*Ro)*d*(Ro+r);
-  system.dC_dy.coeffRef(global_eqn_ids[1], global_var_ids[4]) = 1.0/(Ro*Ro)*eta*(pow(Ro,1.2E+1)*1.0/pow(Ro+r,1.2E+1)*2.0-1.0)*4.0+(Ro*Ro*Ro*Ro)*1.0/pow(Ro+r,7.0)*((Ro*Ro)*W1+(Ro*Ro)*W2+W2*(r*r)+Ro*W2*r*2.0)*2.4E+1-1.0/(Ro*Ro)*W2*(Ro+r)*((Ro*Ro*Ro*Ro*Ro*Ro)*1.0/pow(Ro+r,6.0)*4.0-4.0)*2.0-pow(Ro,1.0E+1)*eta*1.0/pow(Ro+r,1.3E+1)*(Ro*2.0+r*2.0)*4.8E+1;
[…]
