+++ src/solve/debug.h
@@ -40,11 +40,11 @@
  * @brief DEBUG_MSG Macro to print debug messages for debug build
  */
 #ifndef NDEBUG
-#define DEBUG_MSG(str)                                   \
-  do {                                                   \
-    std::cout << "[DEBUG MESSAGE] " << str << std::endl; \
+//#define DEBUG_MSG(str)                                   \
+  //do {                                                   \
+    //std::cout << "[DEBUG MESSAGE] " << str << std::endl; \
   } while (false)
-#else
+//#else
 #define DEBUG_MSG(str) \
   do {                 \
   } while (false)
