### 2025-04-29 b3bcf48
troubleshooting post.py output
