+++ src/model/BloodVesselNew.cpp
@@ -29,23 +29,52 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "BloodVesselNew.h"
+
 #include "Model.h"
 
-
 void BloodVesselNew::setup_dofs(DOFHandler &dofhandler) {
-  Block::setup_dofs_(dofhandler, 7, {"r","v","S","tau","V"});
+  Block::setup_dofs_(dofhandler, 7, {"r", "v", "S", "tau", "V"});
 }
 
+void BloodVesselNew::update_constant(SparseSystem &system,
+                                     std::vector<double> &parameters) {
+  double rho = parameters[global_param_ids[ParamId::rho]];
+  double d = parameters[global_param_ids[ParamId::d]];
+
+  system.E.coeffRef(global_eqn_ids[0], global_var_ids[5]) = d * rho;
+  system.E.coeffRef(global_eqn_ids[2], global_var_ids[7]) = 1.0;
+
+  //
+  system.F.coeffRef(global_eqn_ids[1], global_var_ids[6]) = -1.0;
+  system.F.coeffRef(global_eqn_ids[1], global_var_ids[7]) = 1.0;
+
+  // Qin - Qout - dV/dt = 0
+  system.F.coeffRef(global_eqn_ids[3], global_var_ids[1]) = 1.0;
+  system.F.coeffRef(global_eqn_ids[3], global_var_ids[3]) = -1.0;
+  system.F.coeffRef(global_eqn_ids[3], global_var_ids[8]) = -1.0;
+
+  system.F.coeffRef(global_eqn_ids[4], global_var_ids[8]) = -1.0;
+
+  // dr/dt - v = 0
+  system.F.coeffRef(global_eqn_ids[5], global_var_ids[4]) = 1.0;
+  system.F.coeffRef(global_eqn_ids[5], global_var_ids[5]) = -1.0;
+
+  // Pin - Pout = 0
+  system.F.coeffRef(global_eqn_ids[6], global_var_ids[0]) = 1.0;
+  system.F.coeffRef(global_eqn_ids[6], global_var_ids[2]) = -1.0;
+}
+
+void BloodVesselNew::update_time(SparseSystem &system,
+                                 std::vector<double> &parameters) {
+  get_elastance_values(parameters);
+  system.C(global_eqn_ids[2]) = -a_plus * sigma_o;
+  system.F.coeffRef(global_eqn_ids[2], global_var_ids[7]) = a;
+}
 
 void BloodVesselNew::update_solution(
     SparseSystem &system, std::vector<double> &parameters,
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &y,
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &dy) {
-    
-  get_elastance_values(parameters);  
-  // Get parameters
-  
-  double rho = parameters[global_param_ids[ParamId::rho]];
   double d = parameters[global_param_ids[ParamId::d]];
   double Ro = parameters[global_param_ids[ParamId::Ro]];
   double W1 = parameters[global_param_ids[ParamId::W1]];
@@ -53,69 +82,49 @@ void BloodVesselNew::update_solution(
   double eta = parameters[global_param_ids[ParamId::eta]];
[…]
