+++ src/model/BloodVesselNew.h
@@ -205,7 +205,26 @@ class BloodVesselNew : public Block {
    * @param system System to update contributions at
    * @param parameters Parameters of the model
    */
-  // void update_constant(SparseSystem &system, std::vector<double> &parameters);
+  // void update_constant(SparseSystem &system, std::vector<double>
+  // &parameters);
+
+  /**
+   * @brief Update the constant contributions of the element in a sparse
+   system
+   *
+   * @param system System to update contributions at
+   * @param parameters Parameters of the model
+   */
+  void update_constant(SparseSystem &system, std::vector<double> &parameters);
+
+  /**
+   * @brief Update the time-dependent contributions of the element in a sparse
+   * system
+   *
+   * @param system System to update contributions at
+   * @param parameters Parameters of the model
+   */
+  void update_time(SparseSystem &system, std::vector<double> &parameters);
 
   // /**
   //  * @brief Update the solution-dependent contributions of the element in a
@@ -220,14 +239,15 @@ class BloodVesselNew : public Block {
                        const Eigen::Matrix<double, Eigen::Dynamic, 1> &y,
                        const Eigen::Matrix<double, Eigen::Dynamic, 1> &dy);
 
- /**
+  /**
    * @brief Update the time-dependent contributions of the element in a sparse
    * system
    *
    * @param system System to update contributions at
    * @param parameters Parameters of the model
    */
-  // void update_time(SparseSystem &system, std::vector<double> &parameters, const Eigen::VectorXd &y,
+  // void update_time(SparseSystem &system, std::vector<double> &parameters,
+  // const Eigen::VectorXd &y,
   //   const Eigen::VectorXd &dy);
 
   // private:
@@ -240,10 +260,10 @@ class BloodVesselNew : public Block {
    * @param parameters Parameters of the model
    */
   void get_elastance_values(std::vector<double> &parameters);
-    private:
-      double a = 0.0;   // Chamber Elastance
-      double a_plus = 0.0; // Chamber elastance max
 
+ private:
+  double a = 0.0;       // Chamber Elastance
+  double a_plus = 0.0;  // Chamber elastance max
 
   /**
    * @brief Number of triplets of element
