+++ src/model/Model.h
@@ -45,7 +45,7 @@
 #include "Block.h"
 #include "BlockFactory.h"
 #include "BloodVessel.h"
-#include "BloodVesselNew.h"
+#include "ChamberSphere.h"
 #include "BloodVesselJunction.h"
 #include "ChamberElastanceInductor.h"
 #include "ClosedLoopCoronaryLeftBC.h"
