### 2025-05-05 6d8f950
Update paper.md: minor English edits for the JOSS publication (#173)
