### 2025-05-07 #7595 @menon-karthik
@philipcardiff Thanks for the edits, I have merged your PR.

Also worked through the above bullet points.

The new release is v2.1 here: https://github.com/SimVascular/svZeroDSolver/releases/tag/v2.1

I have archived the release on Zenodo and made sure the archive matches the new release. The archive has DOI [10.5281/zenodo.15359080](https://doi.org/10.5281/zenodo.15359080).
