+++ src/algebra/SparseSystem.cpp
@@ -101,6 +101,7 @@ void SparseSystem::update_jacobian(double time_coeff_ydot,
 void SparseSystem::solve() {
   solver->factorize(jacobian);
   if (solver->info() != Eigen::Success) {
+    std::cout<<"Jacobian: "<<std::endl<<jacobian<<std::endl;
     throw std::runtime_error(
         "System is singular. Check your model (connections, boundary "
         "conditions, parameters).");
