+++ src/model/ParallelRL.h
@@ -1,0 +1,147 @@
+// Copyright (c) Stanford University, The Regents of the University of
+//               California, and others.
+//
+// All Rights Reserved.
+//
+// See Copyright-SimVascular.txt for additional details.
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject
+// to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included
+// in all copies or substantial portions of the Software.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+// TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+// OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+/**
+ * @file ParallelRL.h
+ * @brief model::ParallelRL source file
+ */
+#ifndef SVZERODSOLVER_MODEL_PARALLEL_RL_HPP_
+#define SVZERODSOLVER_MODEL_PARALLEL_RL_HPP_
+
+#include <math.h>
+
+#include "Block.h"
+#include "SparseSystem.h"
+
+/**
+ * @brief Parallel RL circuit element for 0D blood flow modeling
+ *
+ * Models a parallel resistor and inductor with shared pressure at
+ * both the inlet and outlet.
+ *
+ * \f[
+ * \begin{circuitikz} \draw
+ *   node[left] {$Q_{in}$} [-latex] (0,0) -- (0.8,0);
+ *   (1,0) node[anchor=south]{$P_{in}$}
+ *   to [short, -*] (2,0)
+ *   to [R, l=$R$] (5,0)
+ *   to [short, -*] (6,0)
+ *   node[anchor=south]{$P_{out}$}
+ *   (7.2,0) -- (8,0) node[right] {$Q_{out}$};
+ *   (2,0) to (2,-2) to [L, l=$L$] (6,-2) to (6,0);
+ * \end{circuitikz}
+ * \f]
+ *
+ * ### Governing equations
+ *
+ * \f[
+ * P_\text{in}-P_\text{out} = 0 \f]
+ *
+ * \f[
+ * Q_\text{in} - Q_R - Q_L - Q_\text{out} = 0 \f]
+ * 
+ * \f[
[…]
