### 2025-05-09 #7595 @philipcardiff
@menon-karthik : regarding the license, it should be an OSI approved license (see https://joss.readthedocs.io/en/latest/review_criteria.html). Can you clarify which license is used?
