### 2025-05-09 #7595 @philipcardiff
HI again, @menon-karthik. Can you fix the missing and invalid DOIs above? Thanks
