+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{JANSEN2000305,
-title = {A generalized-alpha method for integrating the filtered Navier–Stokes equations with a stabilized finite element method},
+title = {A g}eneralized-alpha method for integrating the filtered Navier–Stokes equations with a stabilized finite element method},
 journal = {Computer Methods in Applied Mechanics and Engineering},
 volume = {190},
 number = {3},
@@ -7,6 +7,7 @@ pages = {305-319},
 year = {2000},
 issn = {0045-7825},
 url = {https://doi.org/10.1016/S0045-7825(00)00203-6},
+doi = {10.1016/S0045-7825(00)00203-6},
 author = {Kenneth E. Jansen and Christian H. Whiting and Gregory M. Hulbert},
 }
 
@@ -70,6 +71,7 @@ url = {https://doi.org/10.1007/s10439-010-0083-6}
   owner            = {pfaller},
   pagetotal        = {406},
   url              = {https://onlinelibrary.wiley.com/doi/book/10.1002/9781118483565},
+  doi              = {10.1002/9781118483565},
 }
 
 @article{pfaller2019importance,
@@ -90,7 +92,8 @@ url = {https://doi.org/10.1007/s10439-010-0083-6}
   volume={16},
   pages={1099--1109},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12265-023-10374-w},
 }
 
 @article{sankaran2012patient,
@@ -136,17 +139,15 @@ url = {https://doi.org/10.1007/s10439-010-0083-6}
   year = {2024}
 }
 @article{richter2024bayesian,
-  title={Bayesian Windkessel calibration using optimized 0D surrogate models}, 
-  author={Jakob Richter and Jonas Nitzler and Luca Pegolotti and Karthik Menon and Jonas Biehler and Wolfgang A. Wall and Daniele E. Schiavazzi and Alison L. Marsden and Martin R. Pfaller},
-  year={2024},
-  volume={2404.14187},
-  journal={arXiv},
-}
-@MISC{eigenweb,
-  author = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
-  title = {Eigen v3},
-  howpublished = {http://eigen.tuxfamily.org},
-  year = {2010}
+  title={Bayesian Windkessel calibration using optimized zero-dimensional surrogate models},
+  author={Richter, Jakob and Nitzler, Jonas and Pegolotti, Luca and Menon, Karthik and Biehler, Jonas and Wall, Wolfgang A and Schiavazzi, Daniele E and Marsden, Alison L and Pfaller, Martin R},
+  journal={Philosophical Transactions A},
+  volume={383},
+  number={2292},
+  pages={20240223},
+  year={2025},
+  publisher={The Royal Society},
+  doi={10.1098/rsta.2024.0223}
 }
 
 @article{Kim2009,
@@ -298,6 +299,7 @@ year = {2019}
       volume={2409.02247},
       journal={arXiv},
       url={https://arxiv.org/abs/2409.02247},
+      doi={10.48550/arXiv.2409.02247}
 }
 
 @article{AFRICA2024109039,
@@ -347,7 +349,7 @@ volume = {135},
 pages = {104641},
 year = {2021},
 issn = {0010-4825},
-doi = {https://doi.org/10.1016/j.compbiomed.2021.104641},
+doi = {10.1016/j.compbiomed.2021.104641},
 url = {https://www.sciencedirect.com/science/article/pii/S0010482521004352},
 author = {F. Regazzoni and A. Quarteroni},
[…]
