+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{JANSEN2000305,
-title = {A g}eneralized-alpha method for integrating the filtered Navier–Stokes equations with a stabilized finite element method},
+title = {A generalized-alpha method for integrating the filtered Navier–Stokes equations with a stabilized finite element method},
 journal = {Computer Methods in Applied Mechanics and Engineering},
 volume = {190},
 number = {3},
