### 2025-05-12 835a294
JOSS review: Fix DOIs (#174)
Co-authored-by: menon-karthik <menon-karthik@users.noreply.github.com>
