### 2025-05-12 #7595 @menon-karthik
@philipcardiff I have updated the title of the archive (https://zenodo.org/records/15359081) and fixed the DOIs. The type of license is a modified BSD-2-Clause license (https://opensource.org/license/bsd-2-clause).
