+++ src/model/FourElementWindkesselBC.cpp
@@ -1,0 +1,69 @@
+// Copyright (c) Stanford University, The Regents of the University of
+//               California, and others.
+//
+// All Rights Reserved.
+//
+// See Copyright-SimVascular.txt for additional details.
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject
+// to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included
+// in all copies or substantial portions of the Software.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+// TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+// OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+#include "FourElementWindkesselBC.h"
+
+void FourElementWindkesselBC::setup_dofs(DOFHandler &dofhandler) {
+  Block::setup_dofs_(dofhandler, 3, {"pressure_c", "pressure_L"});
+}
+
+void FourElementWindkesselBC::update_constant(SparseSystem &system,
+                                            std::vector<double> &parameters) {
+  // P_in - P_L = 0 (constant part)
+  system.F.coeffRef(global_eqn_ids[0], global_var_ids[0]) = 1.0;
+  system.F.coeffRef(global_eqn_ids[0], global_var_ids[3]) = -1.0;
+  
+  // P_L - P_c - R_p*Q_in = 0 (constant part)
+  system.F.coeffRef(global_eqn_ids[1], global_var_ids[2]) = -1.0;
+  system.F.coeffRef(global_eqn_ids[1], global_var_ids[3]) = 1.0;
+  
+  // R_d*Q_in - P_c + P_ref = 0 (constant part)
+  system.F.coeffRef(global_eqn_ids[2], global_var_ids[2]) = -1.0;
+}
+
+void FourElementWindkesselBC::update_time(SparseSystem &system,
+                                        std::vector<double> &parameters) {
+  double inductance = parameters[global_param_ids[0]];
+  double resistance_proximal = parameters[global_param_ids[1]];
+  double capacitance = parameters[global_param_ids[2]];
+  double resistance_distal = parameters[global_param_ids[3]];
[…]
