+++ src/model/FourElementWindkesselBC.h
@@ -1,0 +1,177 @@
+// Copyright (c) Stanford University, The Regents of the University of
+//               California, and others.
+//
+// All Rights Reserved.
+//
+// See Copyright-SimVascular.txt for additional details.
+//
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject
+// to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included
+// in all copies or substantial portions of the Software.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+// TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+// PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+// OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+/**
+ * @file FourElementWindkesselBC.h
+ * @brief model::FourElementWindkesselBC source file
+ */
+#ifndef SVZERODSOLVER_MODEL_FOUR_ELEMENT_WINDKESSELBC_HPP_
+#define SVZERODSOLVER_MODEL_FOUR_ELEMENT_WINDKESSELBC_HPP_
+
+#include "Block.h"
+#include "SparseSystem.h"
+
+/**
+ * @brief Four-Element Windkessel (RDCR) boundary condition.
+ *
+ * Models the mechanical behavior of a Four-Element Windkessel (RDCR) boundary condition.
+ *
+ * \f[
+ * \begin{circuitikz} \draw
+ * node[left] {$Q_{in}$} [-latex] (0,0) -- (0.8,0);
+ * \draw (1,0) node[anchor=south]{$P_{in}$}
+ * to [L, l=$L$, *-] (3,0)
+ * to [R, l=$R_p$, *-] (5,0)
+ * node[anchor=south]{$P_{c}$}
+ * to [R, l=$R_d$, *-*] (7,0)
+ * node[anchor=south]{$P_{ref}$}
+ * (5,0) to [C, l=$C$, *-] (5,-1.5)
+ * node[ground]{};
+ * \end{circuitikz}
+ * \f]
+ *
+ * ### Governing equations
+ *
+ * \f[
+ * P_{in} - P_L - L \dot{Q}_{in} = 0
[…]
