### 2025-05-13 0f0d304
set fixed precision instead of line lengths
