+++ src/solve/csv_writer.cpp
@@ -29,6 +29,7 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "csv_writer.h"
+#include <iomanip>
 
 /**
  * @brief Write results vessel based.
@@ -46,10 +47,9 @@ std::string to_vessel_csv(const std::vector<double> &times,
                           bool mean, bool derivative) {
   // Create string stream to buffer output
   std::stringstream out;
-
-  // Create short and long buffer for lines
-  char sbuff[140];
-  char lbuff[236];
+  
+  // Set floating point format for the entire stream
+  out << std::scientific << std::setprecision(16);
 
   // Write column labels
   if (derivative) {
@@ -110,21 +110,28 @@ std::string to_vessel_csv(const std::vector<double> &times,
         d_outflow_mean /= num_steps;
         d_inpres_mean /= num_steps;
         d_outpres_mean /= num_steps;
-        snprintf(
-            lbuff, 236, "%s,,%.16e,%.16e,%.16e,%.16e,%.16e,%.16e,%.16e,%.16e\n",
-            name.c_str(), inflow_mean, outflow_mean, inpres_mean, outpres_mean,
-            d_inflow_mean, d_outflow_mean, d_inpres_mean, d_outpres_mean);
-        out << lbuff;
+        
+        out << name << ",," 
+            << inflow_mean << "," 
+            << outflow_mean << "," 
+            << inpres_mean << "," 
+            << outpres_mean << "," 
+            << d_inflow_mean << "," 
+            << d_outflow_mean << "," 
+            << d_inpres_mean << "," 
+            << d_outpres_mean << "\n";
       } else {
         for (size_t i = 0; i < num_steps; i++) {
-          snprintf(lbuff, 236,
-                   "%s,%.16e,%.16e,%.16e,%.16e,%.16e,%.16e,%.16e,%.16e,%.16e\n",
-                   name.c_str(), times[i], states[i].y[inflow_dof],
-                   states[i].y[outflow_dof], states[i].y[inpres_dof],
-                   states[i].y[outpres_dof], states[i].ydot[inflow_dof],
-                   states[i].ydot[outflow_dof], states[i].ydot[inpres_dof],
-                   states[i].ydot[outpres_dof]);
-          out << lbuff;
+          out << name << "," 
+              << times[i] << "," 
+              << states[i].y[inflow_dof] << "," 
+              << states[i].y[outflow_dof] << "," 
+              << states[i].y[inpres_dof] << "," 
+              << states[i].y[outpres_dof] << "," 
+              << states[i].ydot[inflow_dof] << "," 
+              << states[i].ydot[outflow_dof] << "," 
+              << states[i].ydot[inpres_dof] << "," 
+              << states[i].ydot[outpres_dof] << "\n";
         }
       }
     } else {
@@ -144,16 +151,20 @@ std::string to_vessel_csv(const std::vector<double> &times,
         outflow_mean /= num_steps;
         inpres_mean /= num_steps;
         outpres_mean /= num_steps;
-        snprintf(sbuff, 140, "%s,,%.16e,%.16e,%.16e,%.16e\n", name.c_str(),
-                 inflow_mean, outflow_mean, inpres_mean, outpres_mean);
-        out << sbuff;
+        
+        out << name << ",," 
+            << inflow_mean << "," 
+            << outflow_mean << "," 
+            << inpres_mean << "," 
+            << outpres_mean << "\n";
       } else {
         for (size_t i = 0; i < num_steps; i++) {
-          snprintf(sbuff, 140, "%s,%.16e,%.16e,%.16e,%.16e,%.16e\n",
[…]
