+++ src/solve/csv_writer.cpp
@@ -29,6 +29,7 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "csv_writer.h"
+
 #include <iomanip>
 
 /**
@@ -47,7 +48,7 @@ std::string to_vessel_csv(const std::vector<double> &times,
                           bool mean, bool derivative) {
   // Create string stream to buffer output
   std::stringstream out;
-  
+
   // Set floating point format for the entire stream
   out << std::scientific << std::setprecision(16);
 
@@ -110,27 +111,19 @@ std::string to_vessel_csv(const std::vector<double> &times,
         d_outflow_mean /= num_steps;
         d_inpres_mean /= num_steps;
         d_outpres_mean /= num_steps;
-        
-        out << name << ",," 
-            << inflow_mean << "," 
-            << outflow_mean << "," 
-            << inpres_mean << "," 
-            << outpres_mean << "," 
-            << d_inflow_mean << "," 
-            << d_outflow_mean << "," 
-            << d_inpres_mean << "," 
-            << d_outpres_mean << "\n";
+
+        out << name << ",," << inflow_mean << "," << outflow_mean << ","
+            << inpres_mean << "," << outpres_mean << "," << d_inflow_mean << ","
+            << d_outflow_mean << "," << d_inpres_mean << "," << d_outpres_mean
+            << "\n";
       } else {
         for (size_t i = 0; i < num_steps; i++) {
-          out << name << "," 
-              << times[i] << "," 
-              << states[i].y[inflow_dof] << "," 
-              << states[i].y[outflow_dof] << "," 
-              << states[i].y[inpres_dof] << "," 
-              << states[i].y[outpres_dof] << "," 
-              << states[i].ydot[inflow_dof] << "," 
-              << states[i].ydot[outflow_dof] << "," 
-              << states[i].ydot[inpres_dof] << "," 
+          out << name << "," << times[i] << "," << states[i].y[inflow_dof]
+              << "," << states[i].y[outflow_dof] << ","
+              << states[i].y[inpres_dof] << "," << states[i].y[outpres_dof]
+              << "," << states[i].ydot[inflow_dof] << ","
+              << states[i].ydot[outflow_dof] << ","
+              << states[i].ydot[inpres_dof] << ","
               << states[i].ydot[outpres_dof] << "\n";
         }
       }
@@ -151,20 +144,15 @@ std::string to_vessel_csv(const std::vector<double> &times,
         outflow_mean /= num_steps;
         inpres_mean /= num_steps;
         outpres_mean /= num_steps;
-        
-        out << name << ",," 
-            << inflow_mean << "," 
-            << outflow_mean << "," 
-            << inpres_mean << "," 
-            << outpres_mean << "\n";
+
+        out << name << ",," << inflow_mean << "," << outflow_mean << ","
+            << inpres_mean << "," << outpres_mean << "\n";
       } else {
         for (size_t i = 0; i < num_steps; i++) {
-          out << name << "," 
-              << times[i] << "," 
-              << states[i].y[inflow_dof] << "," 
-              << states[i].y[outflow_dof] << "," 
-              << states[i].y[inpres_dof] << "," 
-              << states[i].y[outpres_dof] << "\n";
+          out << name << "," << times[i] << "," << states[i].y[inflow_dof]
+              << "," << states[i].y[outflow_dof] << ","
+              << states[i].y[inpres_dof] << "," << states[i].y[outpres_dof]
+              << "\n";
         }
       }
     }
@@ -189,7 +177,7 @@ std::string to_variable_csv(const std::vector<double> &times,
                             const Model &model, bool mean, bool derivative) {
   // Create string stream to buffer output
[…]
