### 2025-05-14 f166cbd
update solver tolerance, recompute reference solutions, relax test tolerance
