### 2025-05-14 bc6be6f
first python jacobian script
