### 2025-05-14 e8f18b8
put definition in yaml file
