+++ scripts/jacobian.py
@@ -1,44 +1,39 @@
-from sympy import symbols, Matrix, diff, simplify, pi
+from sympy import symbols, Matrix, simplify, pi
 from sympy.printing import ccode
 import re
 import pdb
+import yaml
 
-# --- Global Variable Definitions ---
-ny = 9
-y = Matrix(symbols(f'y0:{ny}'))
-dy = Matrix(symbols(f'dy0:{ny}'))
+def load_model(filepath):
+    with open(filepath, 'r') as file:
+        data = yaml.safe_load(file)
 
-# Parameter symbols
-rho, d, Ro, W1, W2, eta, a, a_plus, sigma_o = symbols('rho d Ro W1 W2 eta a a_plus sigma_o')
-time_dependent_symbols = {a, a_plus}
+    variables = Matrix(symbols(' '.join(data['variables'])))
+    derivatives = Matrix(symbols(' '.join(data['derivatives'])))
+    constants = symbols(' '.join(data['constants']))
 
-# Helper functions
-def C(Ro, idx):
-    return (1 + (y[idx] / Ro)) ** 2
+    context = {str(s): s for s in list(variables) + list(derivatives) + list(constants)}
+    context['pi'] = pi
 
-def dC(Ro, idx, d_idx):
-    return diff(C(Ro, idx), y[idx]) * dy[d_idx]
+    exec(data['helper_functions'], context, context)
+    residual_exprs = [eval(res, context, context) for res in data['residuals']]
 
-# --- User-Defined Residuals ---
-residuals = Matrix([
-    rho * d * dy[5] + (d / Ro) * (1 + (y[4] / Ro)) * y[6] - y[2] * C(Ro, 4),
-    -y[6] + y[7] + 4 * (1 - C(Ro, 4) ** -3) * (W1 + C(Ro, 4) * W2) + 2 * eta * dC(Ro, 4, 4) * (1 - 2 * C(Ro, 4) ** -6),
-    dy[7] + a * y[7] - sigma_o * a_plus,
-    y[1] - y[3] - dy[8],
-    4 * pi * Ro ** 2 * C(Ro, 4) * y[5] - dy[8],
-    dy[4] - y[5],
-    y[0] - y[2]
-])
+    residuals = Matrix(residual_exprs)
+    time_dependent = {context[s] for s in data.get('time_dependent', [])}
 
-#
-nr = residuals.shape[0]
-assert ny - 2 == nr, f"Number of residuals ({nr}) should be number of unknowns ({ny}) minus 2."
+    assert time_dependent.issubset({s for s in constants}), "Time dependent variables must be constants"
+    assert len(variables) == len(derivatives), f"Number of variables must be equal to number of derivatives"
+    assert len(variables) - 2 == len(residuals), f"Number of residuals must be number of unknowns minus 2"
 
-def extract_linear_terms(residuals, y, dy):
-    E = Matrix.zeros(nr, ny)
-    F = Matrix.zeros(nr, ny)
+    return variables, derivatives, constants, residuals, time_dependent
+
+def extract_linear(residuals, y, dy):
     def is_constant_coeff(coeff):
         return (coeff and coeff.free_symbols.isdisjoint(y.free_symbols | dy.free_symbols))
+    nr = residuals.shape[0]
+    ny = y.shape[0]
+    E = Matrix.zeros(nr, ny)
+    F = Matrix.zeros(nr, ny)
     for i in range(nr):
         for j in range(ny):
             for mat, dd in zip([E, F], [dy, y]):
@@ -47,25 +42,28 @@ def extract_linear_terms(residuals, y, dy):
                     mat[i, j] = coeff
     return E, F
 
[…]
