### 2025-05-14 ed9a20e
also print constants and variables
