### 2025-05-14 c699bff
reset equations
