+++ src/model/ChamberSphere.cpp
@@ -29,104 +29,109 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "ChamberSphere.h"
+
 #include "Model.h"
 
-
 void ChamberSphere::setup_dofs(DOFHandler &dofhandler) {
-  Block::setup_dofs_(dofhandler, 7, {"r","v","S","tau","V"});
+  Block::setup_dofs_(dofhandler, 7, {"radius", "velo", "stress", "tau", "volume"});
 }
 
+void ChamberSphere::update_constant(SparseSystem &system,
+                                    std::vector<double> &parameters) {
+  const double thick0 = parameters[global_param_ids[ParamId::thick0]];
+  const double rho = parameters[global_param_ids[ParamId::rho]];
+
+  system.E.coeffRef(global_eqn_ids[0], global_var_ids[5]) = rho * thick0;
+  system.F.coeffRef(global_eqn_ids[1], global_var_ids[6]) = -1;
+  system.F.coeffRef(global_eqn_ids[1], global_var_ids[7]) = 1;
+  system.E.coeffRef(global_eqn_ids[2], global_var_ids[8]) = -1;
+  system.E.coeffRef(global_eqn_ids[3], global_var_ids[7]) = 1;
+  system.E.coeffRef(global_eqn_ids[4], global_var_ids[4]) = 1;
+  system.F.coeffRef(global_eqn_ids[4], global_var_ids[5]) = -1;
+  system.F.coeffRef(global_eqn_ids[5], global_var_ids[1]) = 1;
+  system.F.coeffRef(global_eqn_ids[5], global_var_ids[3]) = -1;
+  system.E.coeffRef(global_eqn_ids[5], global_var_ids[8]) = -1;
+  system.F.coeffRef(global_eqn_ids[6], global_var_ids[0]) = 1;
+  system.F.coeffRef(global_eqn_ids[6], global_var_ids[2]) = -1;
+}
+
+void ChamberSphere::update_time(SparseSystem &system,
+                                std::vector<double> &parameters) {
+  get_elastance_values(parameters);
+  system.F.coeffRef(global_eqn_ids[3], global_var_ids[7]) = act;
+}
 
 void ChamberSphere::update_solution(
     SparseSystem &system, std::vector<double> &parameters,
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &y,
     const Eigen::Matrix<double, Eigen::Dynamic, 1> &dy) {
-    
-  get_elastance_values(parameters);  
-  // Get parameters
-  
-  double rho = parameters[global_param_ids[ParamId::rho]];
-  double d = parameters[global_param_ids[ParamId::d]];
-  double Ro = parameters[global_param_ids[ParamId::Ro]];
-  double W1 = parameters[global_param_ids[ParamId::W1]];
-  double W2 = parameters[global_param_ids[ParamId::W2]];
-  double eta = parameters[global_param_ids[ParamId::eta]];
-  double sigma_o = parameters[global_param_ids[ParamId::sigma_o]];
-
-  double Pin = y[global_var_ids[0]]; 
-  double Qin = y[global_var_ids[1]]; 
-  double Pout = y[global_var_ids[2]]; 
-  double Qout = y[global_var_ids[3]]; 
-  double r = y[global_var_ids[4]]; 
[…]
