### 2025-05-14 3c7737b
add more examples and readme
