+++ src/model/BlockType.h
@@ -55,7 +55,7 @@ enum class BlockType {
   closed_loop_heart_pulmonary = 12,
   valve_tanh = 13,
   chamber_elastance_inductor = 14,
-  blood_vessel_new = 15
+  chamber_sphere = 15
 };
 
 /**
