### 2025-05-14 f51041a
tidy up equations
