+++ tests/compute_ref_sol.py
@@ -1,6 +1,8 @@
 import os
 import json
 import pysvzerod
+import argparse
+import sys
 
 this_file_dir = os.path.abspath(os.path.dirname(__file__))
 
@@ -29,7 +31,7 @@ def compute_ref_sol(testname):
     # print for confirmation
     print(f'Reference solution for test case {testname} computed and saved to {result_filename}. Please verify that the results are as expected.')
     
-def compute_all_ref_sol():
+def compute_all_ref_sol(tolerance=None):
     """
     Compute reference solutions for all test cases in the 'cases' directory.
     This function iterates through all JSON files in the 'cases' directory
@@ -38,20 +40,56 @@ def compute_all_ref_sol():
     testfiles = [f for f in os.listdir(os.path.join(this_file_dir, 'cases')) if f.endswith('.json')]
     testfiles.remove("steadyFlow_calibration.json")
     for testfile in testfiles:
+        if tolerance is not None:
+            set_absolute_tolerance(testfile, tolerance)
+
         compute_ref_sol(testfile)
 
+def set_absolute_tolerance(case_name, tolerance):
+    """
+    Set the absolute tolerance for a specific test case.
+
+    Args:
+        case_name: Name of the test case.
+        tolerance: Absolute tolerance value.
+    """
+    # Load the existing configuration
+    with open(os.path.join(this_file_dir, "cases", case_name), "r") as f:
+        config = json.load(f)
+
+    # Set the absolute tolerance
+    config["simulation_parameters"]["absolute_tolerance"] = tolerance
+
+    # Save the updated configuration
+    with open(os.path.join(this_file_dir, "cases", case_name), "w") as f:
+        json.dump(config, f, indent=4)
+        
+
+def main():
+    parser = argparse.ArgumentParser(description="Compute reference solution for a specified test case.")
+    parser.add_argument(
+        "test_case_name",
+        help="Name of the test case or 'all' to compute for all test cases"
+    )
+    parser.add_argument(
+        "--abs-tol",
+        type=float,
+        help="Absolute tolerance to use for the test case (only applies to single test cases)"
+    )
+
+    args = parser.parse_args()
+
+    if args.test_case_name == "all":
+        if args.abs_tol is not None:
+            compute_all_ref_sol(args.abs_tol)
+        else:
+            compute_all_ref_sol()
+    else:
+        print("Computing reference solution for test case:", args.test_case_name)
+        if args.abs_tol is not None:
+            set_absolute_tolerance(args.test_case_name, args.abs_tol)
+        compute_ref_sol(args.test_case_name)
 
 if __name__ == "__main__":
-    # compute the reference solution for a specified test case
-    import sys
-    if len(sys.argv) != 2:
-        print("Usage: python compute_ref_sol.py <test_case_name>")
-        sys.exit(1)
-    else:
-        if sys.argv[1] == "all":
-            compute_all_ref_sol()
-        else:
-            # compute reference solution for a single test case
-            print('computing reference solution for test case:', sys.argv[1])
-            test_case_name = sys.argv[1]
-            compute_ref_sol(test_case_name)
+    
+    main()
