+++ src/model/Model.h
@@ -45,9 +45,9 @@
 #include "Block.h"
 #include "BlockFactory.h"
 #include "BloodVessel.h"
-#include "ChamberSphere.h"
 #include "BloodVesselJunction.h"
 #include "ChamberElastanceInductor.h"
+#include "ChamberSphere.h"
 #include "ClosedLoopCoronaryLeftBC.h"
 #include "ClosedLoopCoronaryRightBC.h"
 #include "ClosedLoopHeartPulmonary.h"
