### 2025-05-16 f2a0ea7
fix reference result i/o issue, fix result matching issue, create diff function.
