+++ tests/utils.py
@@ -17,7 +17,6 @@ this_file_dir = os.path.abspath(os.path.dirname(__file__))
 RTOL_PRES = 1.0e-7
 RTOL_FLOW = 1.0e-7
 
-
 def execute_pysvzerod(testfile, mode):
     """Execute pysvzerod (via Python interface or executable).
 
@@ -51,6 +50,70 @@ def execute_pysvzerod(testfile, mode):
 
     return result, config
 
+def compare_result_with_reference(res, ref, output_variable_based=False):
+    '''
+    Compare the result with the reference.
+
+    Args:
+        res: result as pandas DataFrame 
+        ref: reference result as pandas DataFrame
+        output_variable_based: whether to compare based on columns (True) or row-based with 'name' field (False)
+
+    Returns:
+        pd.DataFrame with columns: ["variable", "name", "relative_difference", "within_tolerance"]
+    '''
+
+    results = []
+
+    if output_variable_based:
+        assert len(res) == len(ref), "Result and reference must have the same number of rows"
+        # merged = ref[["name", "time", "y"]].merge(res[["name", "y"]], on="name", suffixes=("_expected", "_actual"))
+
+        print(res, ref)
+
+        name = ref["name"]
+        y_expected = ref["y"]
+        y_actual = res["y"]
+        tol = RTOL_FLOW if "flow" in name else RTOL_PRES
+        rel_diff = abs(y_actual - y_expected) - tol - (tol * abs(y_expected))
+        within_tol = rel_diff <= 0.0
+
+        for var_name, (rd, wt) in zip(name, zip(rel_diff, within_tol)):
+            # name = res.loc[idx, "name"] if "name" in res.columns else str(idx)
+
+            variable, block_name = var_name.split(":", 1)
+
+            results.append({
+                "variable": variable,
+                "name": block_name,
+                "relative_difference": rd,
+                "within_tolerance": wt
+            })
+
+    else:
+        # Assume 'res' and 'ref' have the same index and column names
+        for col in res.columns:
+            # Determine tolerance based on variable name
+            tol = RTOL_PRES if "pressure" in col else RTOL_FLOW if "flow" in col else None
+            if tol is None:
+                continue
+
+            rel_diff = abs(res[col] - ref[col]) - tol - (tol * abs(ref[col]))
+            within_tol = rel_diff <= 0.0
+
+            # Extract name from index or fallback if index is not informative
+            for idx, (rd, wt) in zip(res.index, zip(rel_diff, within_tol)):
+                name = res.loc[idx, "name"] if "name" in res.columns else str(idx)
+                results.append({
+                    "variable": col,
+                    "name": name,
+                    "relative_difference": rd,
+                    "within_tolerance": wt
+                })
+
+
+    return pd.DataFrame(results, columns=["variable", "name", "relative_difference", "within_tolerance"])
+
 
 def run_with_reference(
         ref,
@@ -58,68 +121,25 @@ def run_with_reference(
         ):
 
 
-    res = pysvzerod.simulate(test_config)
+    res, config = execute_pysvzerod(test_config, "solver")
 
-    if res.shape[1] >= 6:
-        # we have a result with fields [name, time, p_in, p_out, q_in, q_out] SOME HAVE GREATTER LENGTH NEED TO ADDRESS
-        pressure_cols = ["pressure_in", "pressure_out"]
-        flow_cols = ["flow_in", "flow_out"]
+    output_variable_based = config["simulation_parameters"].get("output_variable_based", False)
 
-        bool_df = pd.DataFrame(index=res.index, columns=res.columns, dtype=bool)
+    diff = compare_result_with_reference(res, ref, output_variable_based)
 
[…]
