### 2025-05-16 43d37ca
improve difference error and clean up code
