### 2025-05-16 476041b
print result and reference to debug other operating systems
