### 2025-05-17 0498f0a
test dataframe prints reference and result values
