+++ src/model/FlowReferenceBC.cpp
@@ -1,5 +1,5 @@
-// SPDX-FileCopyrightText: Copyright (c) Stanford University, The Regents of the University of California, and others.
-// SPDX-License-Identifier: BSD-3-Clause
+// SPDX-FileCopyrightText: Copyright (c) Stanford University, The Regents of the
+// University of California, and others. SPDX-License-Identifier: BSD-3-Clause
 #include "FlowReferenceBC.h"
 
 void FlowReferenceBC::setup_dofs(DOFHandler &dofhandler) {
