### 2025-05-18 5afdab6
renamed license to txt for cpack
