### 2025-05-19 5dbb4ab
Rebase recent commits and update license (after resolving merge conflicts) (#180)
Co-authored-by: menon-karthik <menon-karthik@users.noreply.github.com>
Co-authored-by: Martin Pfaller <pfaller@stanford.edu>
Co-authored-by: Nick Dorn <ndorn@stanford.edu>
Co-authored-by: Zachary Sexton <47196674+zasexton@users.noreply.github.com>
Co-authored-by: Philip Cardiff <philipcardiff@gmail.com>
