### 2025-05-19 #7595 @menon-karthik
@philipcardiff  @Kevin-Mattheus-Moerman Thanks for clarifying and forcing us to make the long-overdue change to a standard license!

We have changed the license to BSD-3. I have also merged in some changes from the main branch so it's all up to date. As such, I have made a new release and updated the Zenodo archive.

New release: [v3.0](https://github.com/SimVascular/svZeroDSolver/releases/tag/v3.0)
Updated Zenodo archive: [10.5281/zenodo.15465356](https://doi.org/10.5281/zenodo.15465356)
