+++ paper/paper.bib
@@ -379,14 +379,6 @@ doi = {10.1515/jib-2020-0021},
 year = {2020}
 }
 
-@inproceedings{alastruey2012arterial,
-  title={Arterial pulse wave haemodynamics},
-  author={Alastruey, Jordi and Parker, Kim H and Sherwin, Spencer J},
-  booktitle={11th international conference on pressure surges},
-  pages={401--443},
-  year={2012},
-  organization={Virtual PiE Led t/a BHR Group}
-}
 @article{alastruey2012arterial,
   title={Physical determining factors of the arterial pulse waveform: theoretical analysis and calculation using the 1-D formulation},
   author={Alastruey, Jordi and Passerini, Tiziano and Formaggia, Luca and Peir{\'o}, Joaquim},
