### 2025-05-21 5184a51
clean up time array to prevent rounding
