+++ tests/utils.py
@@ -199,6 +199,7 @@ def get_result(result_array, field, branch, time_step):
 #     and number_of_cardiac_cycles.
 
 #     """
+#     cardiac_cycle_period = 0.99
 
 #     json_paths = [os.path.join(this_file_dir, "cases", "coupledBlock_closedLoopHeart_singleVessel.json"),
 #                   os.path.join(this_file_dir, "cases", "coupledBlock_closedLoopHeart_withCoronaries.json"),]
@@ -213,16 +214,19 @@ def get_result(result_array, field, branch, time_step):
 #         num_pts_per_cycle = sim_params.get("number_of_time_pts_per_cardiac_cycle", 100)
 #         total_pts = num_pts_per_cycle
 
+#         # set cardiac cycle period
+#         data["closed_loop_blocks"][0]["cardiac_cycle_period"] = cardiac_cycle_period
+
 #         # Estimate total time from last point in original t
 #         for block in data.get("external_solver_coupling_blocks", []):
 #             t_original = np.array(block["values"]["t"], dtype=np.float64)
 #             q_original = np.array(block["values"]["Q"], dtype=np.float64)
 
 #             # Create a uniform time grid from t[0] to t[-1]
-#             t_uniform = np.linspace(t_original[0], t_original[-1], total_pts, endpoint=True)
+#             t_uniform = np.linspace(0.0, cardiac_cycle_period, total_pts, endpoint=True)
 
 #             # Interpolate Q to match new time points
-#             interp_q = interp1d(t_original, q_original, kind='linear', fill_value="extrapolate")
+#             interp_q = interp1d(t_uniform, q_original, kind='linear', fill_value="extrapolate")
 #             q_uniform = interp_q(t_uniform)
 
 #             # Update the JSON block
