### 2025-05-21 ccd0e80
increase timesteps and adjust external coupling block array
