### 2025-05-21 00de208
reset back to 100 timesteps
