### 2025-05-22 #7595 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
