### 2025-05-23 #7595 @Kevin-Mattheus-Moerman
@menon-karthik as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. All things seem in order so I will now proceed to process this for acceptance.
